\name{3D_Plot_elem}
\alias{lines3d}
\alias{points3d}
\alias{rect3d}
\alias{text3d}

\title{ 3D Plot Elements}

\description{Add elements to 3D Plot }

\usage{ 
lines3d(x, y, z, pmat = getProjectionMatrix(), ...) \cr
points3d(x, y, z, pmat = getProjectionMatrix(), ...) \cr
rect3d(xrange, yrange, z, pmat = getProjectionMatrix(), ...) \cr
text3d(x, y, z, pmat = getProjectionMatrix(), ...)
}

\arguments{
\item{x}{ }
\item{y}{ }
\item{z}{ }
\item{pmat}{ }
\item{...}{ }
\item{xrange}{}
\item{yrange}{}
}

\details{ }

\value{ }

\references{ }

\author{ RAdamant Development Team \email{team@r-adamant.org}}

\note{ }

\seealso{}

\examples{ }

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
