\name{3D_Plot_pars}
\alias{x.axis3d}
\alias{y.axis3d}
\alias{z.axis3d}
\alias{x.title3d}
\alias{y.title3d}
\alias{z.title3d}


\title{ 3D Plot Axis Formatting}

\description{Add and format labels for 3D Plot }

\usage{ x.axis3d(xlim = NULL, ylim = NULL, zlim = NULL, pmat = getProjectionMatrix(), at = NULL,
labels = NULL, theme.params = getCurrentTheme(), show.labels = TRUE, grid = theme.params[["xgrid"]],
...) \cr
y.axis3d(xlim = NULL, ylim = NULL, zlim = NULL, pmat = getProjectionMatrix(), at = NULL,
labels = NULL, theme.params = getCurrentTheme(), show.labels = TRUE, grid = theme.params[["ygrid"]],
...) \cr
z.axis3d(xlim = NULL, ylim = NULL, zlim = NULL, pmat = getProjectionMatrix(), at = NULL, labels =
NULL, theme.params = getCurrentTheme(), show.labels = TRUE, grid = theme.params[["zgrid"]], ...) \cr

x.title3d(xlim = NULL, ylim = NULL, zlim = NULL, pmat = getProjectionMatrix(), title = "",
theme.params = getCurrentTheme(), ...) \cr
y.title3d(xlim = NULL, ylim = NULL, zlim = NULL, pmat = getProjectionMatrix(), title = "",
theme.params = getCurrentTheme(), ...) \cr
z.title3d(xlim = NULL, ylim = NULL, zlim = NULL, pmat = getProjectionMatrix(), title = "",
theme.params = getCurrentTheme(), ...)
}

\arguments{\item{xlim}{ }\item{ylim}{ }\item{zlim}{ }\item{pmat}{ }\item{at}{ }\item{labels}{
}\item{theme.params}{ }\item{show.labels}{ }\item{grid}{ }\item{...}{ } }

\details{ }

\value{ }

\references{ }

\author{ RAdamant Development Team \email{team@r-adamant.org}}

\note{ }

\seealso{ }

\examples{ }

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
