\name{3dptelem}
\alias{lines3d}
\alias{points3d}
\alias{rect3d}
\alias{text3d}
\title{ 3D Plot Elements}
\description{Add elements to 3D Plot }
\usage{ 
lines3d(x, y, z, pmat = getProjectionMatrix(), ...) 
points3d(x, y, z, pmat = getProjectionMatrix(), ...) 
rect3d(xrange, yrange, z, pmat = getProjectionMatrix(), ...) 
text3d(x, y, z, pmat = getProjectionMatrix(), ...)
}
\arguments{
\item{x}{ X axis}
\item{y}{ Y axis}
\item{z}{ Z axis}
\item{pmat}{ pamt}
\item{...}{ Further arguments to or from other methods}
\item{xrange}{ xrange}
\item{yrange}{ yrange}
}
\author{ RAdamant Development Team \email{team@r-adamant.org}}


