\name{Abi}
\alias{Abi}

\title{Absolute Breath Index - ABI}

\description{Compute Absolute Breath Index (Technical Analysis)}

\usage{(X, lag = 5)}

\arguments{ 
\item{X}{ Input numerical series}
\item{lag}{ Number of lags}
\item{plot}{ LOGICAL. Return plot.}
}


\details{The absolute breadth index (ABI) measures the amount of volatile on the New York Stock
Exchange. The indicator is truly a momentum indicator since it only tracks the movement of the
issues on the exchange and not their direction. The overall purpose of the indicator is to determine
when the market is most volatile, thus providing the greatest opportunity for larger profits due to
increased volatility. \cr Calculation: \cr \deqn{ABI = |Advancing \ Issues - Declining \ issues|}}

\value{empty}

\references{empty}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{empty}

\examples{empty}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
