\name{abs_avdi}
\alias{abs_avdi}

\title{Absolute average discard}

\description{Calculate Absolute average discard for model evaluation }

\usage{abs_avdi(actual, predicted, pc = FALSE) }

\arguments{ \item{actual}{ }\item{predicted}{ }\item{pc}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
