\name{absavdi}
\alias{abs_avdi}
\title{Absolute average discard}
\description{Calculate Absolute average discard for model evaluation }
\usage{abs_avdi(actual, predicted, pc = FALSE) }
\arguments{\item{actual}{ actual}\item{predicted}{ predicted}\item{pc}{ pc}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
