\name{Accuracy}

\alias{accuracy}
\alias{accuracy.scorecard}

\title{Accuracy}

\description{Get accuracy measure from the results of a classification model.}

\usage{
accuracy(x, ...) \cr
accuracy.scorecard(x, th) }

\arguments{
\item{x}{ }
\item{th}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
