\name{empty}
\alias{empty}

\title{empty}

\description{empty}

\usage{empty}

\arguments{ empty}

\details{Acceleration/Deceleration Technical Indicator (AC) measures acceleration and deceleration of the current driving force.  \cr This indicator will change direction before any changes in the driving force, which, it its turn, will change its direction before the price.  \cr The nought line is basically the spot where the driving force is at balance with the acceleration.  \cr If Acceleration/Deceleration is higher than nought, then it is usually easier for the acceleration to continue the upward movement (and vice versa in cases when it is below nought). }

\value{empty}

\references{empty}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{empty}

\examples{empty}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
