\name{adi}
\alias{ADind}
\title{Advance-Decline Indicator}
\description{Advance-Decline Indicator (Technical Analysis)}
\usage{ADind(close, high, low, lag = 5)}
\arguments{
\item{close}{ VECTOR. Close price.}
\item{high}{ VECTOR. high price.}
\item{low}{ VECTOR. Low price.}
\item{lag}{ INTEGER. Number of lag periods.}
}
\details{A technical analysis tool that represents the total difference
between the number of advancing and declining security prices. \cr The
advance/decline index can provide much more insight into the movements of the
market. \cr In general, rising values of the advance/decline can be used to
confirm the likelihood that an upward trend will continue. If the market is up
but there are more declining issues than advancing ones, it is usually a sign
that the market is losing its breadth and may be getting ready to change
direction.}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
