\name{empty}
\alias{empty}

\title{empty}

\description{empty}

\usage{empty}

\arguments{ empty}

\details{A technical analysis tool that represents the total difference between the number of advancing and declining security prices. \cr The advance/decline index can provide much more insight into the movements of the market. \cr In general, rising values of the advance/decline can be used to confirm the likelihood that an upward trend will continue. If the market is up but there are more declining issues than advancing ones, it's usually a sign that the market is losing its breadth and may be getting ready to change direction.}

\value{empty}

\references{empty}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{empty}

\examples{empty}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
