\name{ADrat}
\alias{ADrat}

\title{Advance Decline ratio}

\description{Compute Advance Decline ratio (Technical Analysis)}

\usage{ADrat(close, high, low, lag) }

\arguments{ \item{close}{ }\item{high}{ }\item{low}{ }\item{lag}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
