\name{empty}
\alias{empty}

\title{empty}

\description{empty}

\usage{empty}

\arguments{ empty}

\details{The Advance Decline Ratio is also market breadth indicator. It is calculated by dividing the number of advancing issues by the number of declining issues using daily or weekly NYSE data. \cr It works very well as an overbought-oversold indicator and as a momentum indicator. A moving average should be used to smooth out the swings.  \cr Rules: \itemize{ \item If the value is higher than 1, that means the more issues are presently advancing than declining. \item Values between 0 and 1 indicate that more issues are currently declining in price.} Calculation:  \cr \deqn{ ADratio = \frac{Advancing \: Issues}{Declining \: Issues}} }

\value{empty}

\references{empty}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{empty}

\examples{empty}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
