\name{ama}
\alias{ama}
\title{General Adaptive Moving Average}
\description{General Adaptive Moving Average, computed on each column of the
input data X.}
\usage{
ama(X
    , ar.ord = 1
    , ma.ord = 1
    , func = NULL
    , padding = 0
    , type = "AMA"
    , plot = FALSE
    , ...
    ) 
}
\arguments{ 
\item{X}{ X}
\item{ ar.ord}{ ar.ord}
\item{ ma.ord}{ ma.ord}
\item{ func}{func}
\item{ padding}{padding}
\item{ type}{ type}
\item{ plot}{ LOGICAL. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}


