\name{apply.format}
\alias{apply.format}

\title{Apply number format}

\description{Plotting tools}

\usage{apply.format(x, fmt = NULL) }

\arguments{ \item{x}{ }\item{fmt}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
