\name{archlm}

\alias{Archlm}

\title{ARCH-LM test}

\description{Compute ARCH-LM test}

\usage{Archlm(x, lags, std=FALSE, plot.acf=FALSE) }

\arguments{ 
\item{x}{ x}
\item{lags}{ lags}
\item{std}{ std}
\item{plot.acf}{ plot.acf}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
