\name{archlm.test}
\alias{archlm.test}

\title{ARCH-LM test}

\description{Compute ARCH-LM test}

\usage{archlm.test(X, lags, order = c(1, 0, 1)) }

\arguments{ \item{X}{ }\item{lags}{ }\item{order}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
