\name{Arma.Spec}
\alias{Arma.Spec}

\title{Arma spectral representation}

\description{Spectral representation based on ARMA models}

\usage{Arma.Spec(X, ar_ord = 1, ma_ord = 1, vfreq = NULL) }

\arguments{ \item{X}{ }\item{ar_ord}{ }\item{ma_ord}{ }\item{vfreq}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
