\name{armaspc}
\alias{Arma.Spec}
\title{Arma spectral representation}
\description{Spectral representation based on ARMA models}
\usage{Arma.Spec(X, ar_ord = 1, ma_ord = 1, vfreq = NULL) }
\arguments{\item{X}{ X}\item{ar_ord}{ ar_ord}\item{ma_ord}{ ma_ord}\item{vfreq}{ vfreq}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


