\name{Arms}
\alias{Arms}

\title{Arms index}

\description{Compute Arms index (Technical Analysis)}

\usage{Arms(X, Volume, lag, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{Volume}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
