\name{arodown}
\alias{arodown}
\title{Aroon Down oscillator}
\description{Compute Aroon Down oscillator (Technical Analysis)}
\usage{arodown(X, lag = 5, plot = TRUE, ...) }
\arguments{
\item{X}{ X}
\item{ lag}{ Integer. Number of lag periods.}
\item{ plot}{ Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


