\name{aroon}
\alias{aroon}

\title{Aroon oscillator}

\description{Compute Aroon oscillator (Technical Analysis)}

\usage{aroon(X, lag = 5, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
