\name{aroud}
\alias{aroud}
\title{Aroon Down oscillator}
\description{Compute Aroon Down oscillator (Technical Analysis)}
\usage{aroud(X, lag = 5, plot = TRUE, ...) }
\arguments{
\item{X}{ X}
\item{ lag}{ lag}
\item{ plot}{ plot}
\item{ ...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


