\name{as.fs}
\alias{as.fs}

\title{Convert Yahoo! Data into Financial Series object}

\description{}

\usage{as.fs(X, SName = "", Symbol = "") }

\arguments{ \item{X}{ }\item{SName}{ }\item{Symbol}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
