\name{av_er}
\alias{av_er}

\title{Average error}

\description{Calculate Average error for model evaluation }

\usage{av_er(actual, predicted, pc = FALSE) }

\arguments{ \item{actual}{ }\item{predicted}{ }\item{pc}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
