\name{aver}
\alias{av_er}
\title{Average error}
\description{Calculate Average error for model evaluation }
\usage{av_er(actual, predicted, pc = FALSE) }
\arguments{\item{actual}{ actual}\item{predicted}{ predicted}\item{pc}{ pc}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
