\name{BinCoef}
\alias{BinCoef}

\title{Binomial coefficient}

\description{Calculate binomial coefficient}

\usage{BinCoef(N, n) }

\arguments{ \item{N}{ }\item{n}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
