\name{Bol.Fib}
\alias{Bol.Fib}

\title{Bollinger Bands – Fibonacci ratio}

\description{Compute Bollinger Bands – Fibonacci ratio (Technical Analysis)}

\usage{Bol.Fib(Close, High, Low, win.size = 5, fibo = c(1.618, 2.618, 4.236), plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{win.size}{ }\item{fibo}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
