\name{BolBandB}
\alias{BolBandB}

\title{Bollinger Bands B}

\description{Compute Bollinger Bands B (Technical Analysis)}

\usage{BolBandB(Close, High, Low, fact = 2, win.size = 5, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{fact}{ }\item{win.size}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
