\name{bolfib}
\alias{Bol.Fib}
\title{Bollinger Bands - Fibonacci ratio}
\description{Compute Bollinger Bands - Fibonacci ratio (Technical Analysis)}
\usage{
Bol.Fib(Close, High, Low
    , win.size = 5
    , fibo = c(1.618, 2.618, 4.236)
    , plot = FALSE
    , ...
    ) 
}
\arguments{
\item{Close}{ Vector. Close price.}
\item{   High}{ Vector. High price.}
\item{   Low}{ Vector. Low price.}
\item{   win.size}{ win.size}
\item{   fibo}{ fibo}
\item{   plot}{ Logical. If TRUE plot is returned.}
\item{   ...}{ Further arguments to or from other methods.}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


