\name{boot}
\alias{boot}

\title{General bootstrapping function}

\description{}

\usage{boot(X, nboots = 100, func = NULL, init = NULL, message = "Bootstrapping...", ...) }

\arguments{ \item{X}{ }\item{nboots}{ }\item{func}{ }\item{init}{ }\item{message}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
