\name{bop}
\alias{Bop}
\title{Balance of Power}
\description{Compute Balance of Power (Technical Analysis)}
\usage{Bop(Close, Open, High, Low, smoothed = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Open}{ VECTOR. Open price.}\item{High}{ VECTOR.
High price.}\item{Low}{ VECTOR. Low price.}\item{smoothed}{ smoothed}\item{...}{ Further arguments
to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


