\name{box3d}
\alias{box3d}

\title{3D box}

\description{Plotting tools}

\usage{box3d(x, y, z, pmat = getProjectionMatrix(), half = FALSE, ...) }

\arguments{ \item{x}{ }\item{y}{ }\item{z}{ }\item{pmat}{ }\item{half}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
