\name{breadth}
\alias{Breadth}
\title{Breadth trusth indicator}
\description{Compute Breadth trusth indicator (Technical Analysis)}
\usage{Breadth(X, lag = 5, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}




\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


