\name{BroMot2D}
\alias{BroMot2D}

\title{2-dimensional Browniam motion}

\description{Simulate a bi-dimensional standard Brownian motion }

\usage{BroMot2D(nsim, T, S0, mi, sigma, geom = TRUE, same.rnd = FALSE, laydisp = NULL, plot = TRUE, ...) }

\arguments{ \item{nsim}{ }\item{T}{ }\item{S0}{ }\item{mi}{ }\item{sigma}{ }\item{geom}{ }\item{same.rnd}{ }\item{laydisp}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
