\name{BS.formula}
\alias{BS.formula}

\title{Black & Scholes formula}

\description{Black & Scholes analytical formula}

\usage{BS.formula(type = c("call", "put")) }

\arguments{ \item{type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
