\name{BS.greeks}
\alias{BS.greeks}

\title{Black & Scholes greeks}

\description{Calculate analytically Black & Scholes greeks}

\usage{BS.greeks(X = NULL, ...) }

\arguments{ \item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
