\name{BS.ImpVol}
\alias{BS.ImpVol}

\title{Black & Scholes Implied volatility}

\description{Calculate Black & Scholes Implied volatility}

\usage{BS.ImpVol(P, under, strike, rfr, sigma, maty, yield, interval = c(-20, 20), calc.type =
c("standard", "lognorm", "gammarec"), opt.type = c("call", "put")) }

\arguments{ \item{P}{ }\item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }\item{interval}{ }\item{calc.type}{ }\item{opt.type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
