\name{BS.moments}
\alias{BS.moments}

\title{Black & Scholes moments}

\description{Calculate first four moments for Black & Scholes}

\usage{BS.moments(BS = NULL, under, rfr, sigma, yield, maty) }

\arguments{ \item{BS}{ }\item{under}{ }\item{rfr}{ }\item{sigma}{ }\item{yield}{ }\item{maty}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
