\name{BS.price.default}
\alias{BS.price.default}

\title{Black & Scholes price generic}

\description{Generic method for Black & Scholes price}

\usage{BS.price.default(under, strike, rfr, sigma, maty, yield, calc.type = c("standard", "lognorm",
"gammarec"), opt.type = c("call", "put")) }

\arguments{ \item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }\item{calc.type}{ }\item{opt.type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
