\name{BS.price.gamr}
\alias{BS.price.gamr}

\title{Black & Scholes price – Gamma reciprocal}

\description{Calculate option price (Call / Put) with Gamma Reciprocal Black & Scholes}

\usage{BS.price.gamr(under, strike, rfr, sigma, maty, yield, opt.type = c("call", "put")) }

\arguments{ \item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }\item{opt.type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
