\name{BS.price}
\alias{BS.price}

\title{Black & Scholes price (generic method)}

\description{Generic method for Black & Scholes price}

\usage{BS.price(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
