\name{BS.price.std}
\alias{BS.price.std}

\title{Black & Scholes price – Standard}

\description{Calculate option price (Call / Put) with Standard Black & Scholes price}

\usage{BS.price.std(under, strike, rfr, sigma, maty, yield, opt.type = c("call", "put")) }

\arguments{ \item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }\item{opt.type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
