\name{bsgreeks}
\alias{BS.greeks}
\title{Black & Scholes greeks}
\description{Calculate analytically Black & Scholes greeks}
\usage{BS.greeks(X = NULL, ...) }
\arguments{ \item{X}{ X}\item{...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
