\name{BS.ImpVol}
\alias{BS.ImpVol}
\title{Black & Scholes Implied volatility}
\description{Calculate Black & Scholes Implied volatility}

\usage{
BS.ImpVol(P, under, strike, rfr, sigma, maty, 
yield, interval = c(-20, 20), 
calc.type =c("standard", "lognorm", "gammarec"), 
opt.type = c("call", "put")) 
}

\arguments{\item{P}{ P}\item{under}{ under}\item{strike}{ strike}\item{rfr}{ rfr}\item{sigma}{
sigma}\item{maty}{ maty}\item{yield}{ yield}\item{interval}{ interval}\item{calc.type}{
calc.type}\item{opt.type}{ opt.type}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
