\name{buypre}
\alias{buypre}
\title{Buying pressure indicator}
\description{Compute Buying pressure indicator (Technical Analysis)}
\usage{buypre(Close, Low, lag = 5, plot = FALSE, ...) }
\arguments{
\item{Close}{ Vector. Close price.}
\item{   Low}{ Vector. Low price.}
\item{   lag}{ Integer. Number of lag periods.}
\item{   plot}{ Logical. If TRUE plot is returned.}
\item{   ...}{ Further arguments to or from other methods.}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


