\name{CalcPairs}

\alias{CalcPairs}

\title{Calculated Pairs}

\description{Calculated the number of pairs based on the results of a classification model.}

\usage{CalcPairs(target, pred, segm_fact = 0.002) }

\arguments{
\item{target}{ }
\item{pred}{ }
\item{segm_fact}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
