\name{calcprs}
\alias{CalcPairs}
\title{Calculated Pairs}
\description{Calculated the number of pairs based on the results of a classification model.}
\usage{CalcPairs(target, pred, segm_fact = 0.002) }
\arguments{\item{target}{ target}\item{pred}{ pred}\item{segm_fact}{ segm_fact}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
