\name{CallStackLevels}
\alias{CallStackLevels}

\title{Logging utilities}

\description{Tool for Logging operations}

\usage{CallStackLevels(maxLev = 100) }

\arguments{ \item{maxLev}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
