\name{Capm.default}
\alias{Capm.default}

\title{Capm – default method}

\description{Default method for CAPM}

\usage{Capm.default(PTF, PTF_M, rf = NULL, rfr = NULL, ...) }

\arguments{ \item{PTF}{ }\item{PTF_M}{ }\item{rf}{ }\item{rfr}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
