\name{cci.v2}
\alias{cci.v2}

\title{Commodity channel index v02}

\description{Compute Commodity channel index v02 (Technical Analysis)}

\usage{cci.v2(High, Low, Close, lag = 5, plot = FALSE, ...) }

\arguments{ \item{High}{ }\item{Low}{ }\item{Close}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
