\name{cci.v2}
\alias{cci.v2}
\title{Commodity channel index v02}
\description{Compute Commodity channel index v02 (Technical Analysis)}
\usage{cci.v2(High, Low, Close, lag = 5, plot = FALSE, ...) }
\arguments{\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR. Low price.}\item{Close}{ VECTOR.
Close price.}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
