\name{Ch.vol}
\alias{Ch.vol}

\title{Chaikin volatility indicator}

\description{Compute Chaikin volatility indicator (Technical Analysis)}

\usage{Ch.vol(High, Low, Close, lag = 5, plot = FALSE, ...) }

\arguments{ \item{High}{ }\item{Low}{ }\item{Close}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
