\name{chaosAcc}
\alias{chaosAcc}

\title{Chaos Accelerator oscillator}

\description{Compute Chaos Accelerator oscillator (Technical Analysis)}

\usage{chaosAcc(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
