\name{chist}
\alias{chist}

\title{Custom histogram function}

\description{}

\usage{chist(x, nclass = min(max(round(NROW(x)/10), 10), NROW(x)), density = c("kernel", "normal"),
kernel = c("gaussian", "epanechnikov", "rectangular", "triangular", "biweight", "cosine",
"optcosine"), theme.params = getCurrentTheme(), main = "Histogram and Kernel Density Estimation",
xtitle = NULL, ytitle = NULL, legend = NULL, show.legend = TRUE, normalised = FALSE, ...) }

\arguments{ \item{x}{ }\item{nclass}{ }\item{density}{ }\item{kernel}{ }\item{theme.params}{ }\item{main}{ }\item{xtitle}{ }\item{ytitle}{ }\item{legend}{ }\item{show.legend}{ }\item{normalised}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
