\name{chvol}
\alias{Ch.vol}
\title{Chaikin volatility indicator}
\description{Compute Chaikin volatility indicator (Technical Analysis)}
\usage{
Ch.vol(High, Low, Close, lag = 5, plot = FALSE, ...) 
}

\arguments{
\item{High}{ Vector. High price.}
\item{   Low}{ Vector. Low price.}
\item{   Close}{ Vector. Close price.}
\item{   lag}{ Integer. Number of lag periods.}
\item{   plot}{ Logical. If TRUE plot is returned.}
\item{   ...}{ Further arguments to or from other methods.}}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


