\name{Clust}
\alias{Clust}

\title{Time series clusters}

\description{Create a “simple” cluster partition of a time series}

\usage{Clust(X, y = NULL, n_clust = 5, bk.type = c("quantile", "uniform", "custom"), custom_breaks = NULL, lab.dig = 0) }

\arguments{ \item{X}{ }\item{y}{ }\item{n_clust}{ }\item{bk.type}{ }\item{custom_breaks}{ }\item{lab.dig}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
