\name{clv}
\alias{clv}
\title{Close Location value oscillator}
\description{Compute Close Location value oscillator (Technical Analysis)}
\usage{clv(Close, High = NULL, Low = NULL, plot = TRUE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


