\name{cmf}
\alias{cmf}
\title{Chaikin Money Flow}
\description{Compute Chaikin Money Flow (Technical Analysis)}
\usage{cmf(Close, Low, High, Volume, plot = FALSE, ...) }

\arguments{ \item{Close}{ VECTOR. Close price.}\item{Low}{ VECTOR. Low price.}\item{High}{ VECTOR.
High
price.}\item{Volume}{ Volume}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further
arguments to or from other
methods}}




\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


