\name{cmof}
\alias{cmof}

\title{Chande Momentum Oscillator}

\description{Compute Chande Momentum Oscillator (Technical Analysis)}

\usage{cmof(X, lag = 5, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
