\name{colin.pairs}
\alias{colin.pairs}
\title{Co-Linearity analysis}
\description{This function performs a Co-Linearity analysis between the columns of X: Correlation
factors between columns are computed,and pairs of columns with a correlation factor higher than a
specified threshold are returned.}
\usage{colin.pairs(X, trsh = 0.8) }
\arguments{ \item{X}{ X}\item{trsh}{ trsh}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
