\name{colinred}
\alias{colin.reduce}
\title{Co-Linearity reduction}
\description{Perform a cross Co-Linearity analysis between the columns of Y and X, and for each Yi
returns a reduced set of columns of X obtained after removing those columns of X that are too
correlated (one for each co-linear pair). In the removal process, those columns of X that are most
correlated to Yi are kept.}
\usage{colin.reduce(Y, X, max.iter = 100, trsh = 0.85) }
\arguments{\item{Y}{ Y}\item{X}{ X}\item{max.iter}{ max.iter}\item{trsh}{ trsh}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
