\name{combine.fs}
\alias{combine.fs}

\title{Combine Multiple Financial Series}

\description{}

\usage{combine.fs(..., which = "Close") }

\arguments{ \item{...}{ }\item{which}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
