\name{comma.Fmt}
\alias{comma.Fmt}

\title{Comma style formatting}

\description{Plotting tools}

\usage{comma.Fmt(x, digits = 3, ...) }

\arguments{ \item{x}{ }\item{digits}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
