\name{confmat}
\alias{confusionM.default}
\alias{confusionM}
\alias{confusionM.scorecard}
\title{Confusion matrix}
\description{Build confusion matrix based on the results of a classification model.}

\usage{
confusionM(target, ...)
\method{confusionM}{default}(target, pred, th, ...) 
\method{confusionM}{scorecard}(target, th, ...)
}

\arguments{
\item{target}{ target}
\item{pred}{ pred}
\item{th}{ th}
\item{...}{ Further arguments to or from other methods}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
