\name{confusionMat}

\alias{confusionM.default}
\alias{confusionM}
\alias{confusionM.scorecard}


\title{Confusion matrix}

\description{Build confusion matrix based on the results of a classification model.}

\usage{
confusionM.default(orig, pred, th) \cr
confusionM.scorecard(x, th) 
}

\arguments{
\item{x}{} 
\item{orig}{ }
\item{pred}{ }
\item{th}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
