\name{Cplot3d}
\alias{cplot3d}

\title{ 3-Dimensional plotting}

\description{Workhorse function for 3D automatic plotting}

\usage{cplot3d(x, y, z, ...)}

\arguments{ 
\item{x}{ x coordinates for the plot}
\item{y}{ y coordinates for the plot}
\item{z}{ z coordinates for the plot}
\item{fill}{ }
\item{main}{}
\item{xtitle}{ }
\item{ytitle}{ }
\item{ztitle}{ }
\item{xlim}{ }
\item{ylim}{}
\item{zlim}{ }
\item{xlabels}{ }
\item{ylabels}{ }
\item{zlabels}{ }
\item{pre}{}
\item{post}{ }
\item{theme.params}{ }
\item{overrides}{ }
\item{new.device}{}
\item{append}{ }
\item{axis}{ }
\item{show.labels}{ }
\item{show.xlabels}{}
\item{show.ylabels}{ }
\item{show.zlabels}{ }
\item{...}{ }
}

\details{}

\value{}

\references{}

\author{ RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
