\name{cramv}
\alias{cramv}

\title{Cramers V}

\description{Calculate Cramers V}

\usage{cramv(x, y) }

\arguments{ \item{x}{ }\item{y}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
