\name{create.empty.plot}
\alias{create.empty.plot}

\title{Create empty plot}

\description{Plotting tools}

\usage{create.empty.plot(X, base = NULL, new.device = FALSE, main = "", theme.params =
getCurrentTheme(), two.sides = FALSE, set.margins = TRUE, ...) }

\arguments{ \item{X}{ }\item{base}{ }\item{new.device}{ }\item{main}{ }\item{theme.params}{ }\item{two.sides}{ }\item{set.margins}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
