\name{cross.colin}
\alias{cross.colin}

\title{Cross collinearity}

\description{Perform a cross Co-Linearity analysis between the columns of Y and X: Correlation factors between each column Yi and all columns of X are calculated for different time lags.Also pairs of columns of X with a correlation factor higher than a specified threshold are returned.}

\usage{cross.colin(Y, X, max.lag = 8, trsh = 0.8) }

\arguments{ \item{Y}{ }\item{X}{ }\item{max.lag}{ }\item{trsh}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
