\name{CRR.BinTree}
\alias{CRR.BinTree}

\title{CRR Binomial Tree}

\description{Option evaluation with Cox, Rossand and Rubinstein Binomial Tree}

\usage{CRR.BinTree(Nsteps, under, strike, rfr, sigma, maty, yield, life, ret.steps = FALSE) }

\arguments{ \item{Nsteps}{ }\item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }\item{life}{ }\item{ret.steps}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
