\name{dataset}

\alias{ex_ptf}
\alias{ex_ts}
\alias{ex_fs}
\alias{ex_credit}

\docType{data}

\title{Example datasets for portfolio and time series analysis}

\description{
\code{ex_ts}: Univariate timeseris of 126 observations; 
\code{ex_ptf}: Matrix of returns: 60 rows and 8 colums. The first column is taken as a "market fund"
and the other 7 columns are 8 possible indexes.
\code{ex_fs}: An object of class "fs" containing financial series: 252 rows and 6 colums. 
}

\usage{
data(ex_ts)
data(ex_ptf)
data(ex_fs)
data(ex_credit)
}


\source{Artificially created.}


