\name{decimals}
\alias{decimals}

\title{Count decimal}

\description{Count decimal}

\usage{decimals(x, max.digits = 10, ...) }

\arguments{ \item{x}{ }\item{max.digits}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
