\name{demark}
\alias{demark}

\title{DeMark indicator}

\description{Compute DeMark indicator (Technical Analysis)}

\usage{demark(High, Low, Close, lag = 5, plot = FALSE, ...) }

\arguments{ \item{High}{ }\item{Low}{ }\item{Close}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
