\name{dgpd}
\alias{dgpd}

\title{Generalised Pareto Distribution (GPD) - Density function}

\description{}

\usage{dgpd(X, xi = 0.1, sigma = 1, trsh = 0) }

\arguments{ \item{X}{ }\item{xi}{ }\item{sigma}{ }\item{trsh}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
