\name{dpo}
\alias{dpo}
\title{Detrended price oscillator}
\description{Compute Detrended price oscillator (Technical Analysis)}
\usage{dpo(Close, lag = 5, plot = TRUE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{lag}{Integer. Number of lag
periods.}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from
other methods.}}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


