\name{drawdown.default}
\alias{drawdown.default}

\title{Drawdown – default method}

\description{}

\usage{drawdown.default(x, FUN = max, relative = FALSE, plot = FALSE, ...) }

\arguments{ \item{x}{ }\item{FUN}{ }\item{relative}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
