\name{drawdown}
\alias{drawdown}

\title{Drawdown – method}

\description{}

\usage{drawdown(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
