\name{EdgeFact}
\alias{EdgeFact}

\title{Edge Factor (B&S)}

\description{Edgeworth adaption factors}

\usage{EdgeFact(x, s, k) }

\arguments{ \item{x}{ }\item{s}{ }\item{k}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
