\name{Edgeworth.price}
\alias{Edgeworth.price}

\title{Edgeworth option price}

\description{Option evaluation with Edgeworth adapted Binomial Tree}

\usage{Edgeworth.price(init, p, under, strike, rfr, sigma, maty, yield) }

\arguments{ \item{init}{ }\item{p}{ }\item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
