\name{EdgeWorthDist}
\alias{EdgeWorthDist}

\title{Edgeworth distribution}

\description{Simulate empirical Edgeworth distribution}

\usage{EdgeWorthDist(init, Nsteps) }

\arguments{ \item{init}{ }\item{Nsteps}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
