\name{edwdist}
\alias{EdgeWorthDist}
\title{Edgeworth distribution}
\description{Simulate empirical Edgeworth distribution}
\usage{EdgeWorthDist(init, Nsteps, p=0.5) }
\arguments{\item{init}{ init}\item{Nsteps}{ Nsteps}\item{p}{ p}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


