\name{edwprice}
\alias{Edgeworth.price}
\title{Edgeworth option price}
\description{Option evaluation with Edgeworth adapted Binomial Tree}
\usage{Edgeworth.price(init, under, strike, rfr, sigma, maty, yield) }
\arguments{ \item{init}{ init}\item{under}{ under}\item{strike}{ strike}\item{rfr}{
rfr}\item{sigma}{ sigma}
\item{maty}{ maty}\item{yield}{ yiels}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


