\name{eom}
\alias{eom}

\title{Ease of Movement oscillator}

\description{Compute Ease of Movement oscillator (Technical Analysis)}

\usage{eom(Close, High = NULL, Low = NULL, Vol = NULL, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{Vol}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
