\name{erf}
\alias{erf}

\title{Elder Ray force }

\description{Compute Elder Ray force  (Technical Analysis)}

\usage{erf(Close, High = NULL, Low = NULL, lag = 13, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
