\name{extrBreak}

\alias{extrBreak}

\title{Extract breaks}

\description{extract specificied break from an object of class Factorise}

\usage{extrBreak(var) }

\arguments{\item{var}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
