\name{extrbrk}
\alias{extrBreak}
\title{Extract breaks}
\description{extract specificied break from an object of class Factorise}

\usage{extrBreak(var, Factors) }

\arguments{\item{var}{ var}\item{Factors}{ Factors}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
