\name{extrdd}
\alias{ExtremeDD}
\title{Maximum / Minimum drawdown}
\description{Calculate Mximum / Minimum DrawDown}
\usage{ExtremeDD(DD, FUN, lag = 1, rolling = FALSE, plot = TRUE, ...) }
\arguments{\item{DD}{ OBJECT of class "drawdown"}\item{FUN}{ FUN}\item{lag}{ INTEGER. Number of lag
periods.}\item{rolling}{ rolling}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


