\name{ExtremeDD}
\alias{ExtremeDD}

\title{Maximum / Minimum drawdown}

\description{}

\usage{ExtremeDD(DD, FUN, lag = 1, rolling = FALSE, plot = TRUE, ...) }

\arguments{ \item{DD}{ }\item{FUN}{ }\item{lag}{ }\item{rolling}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
