\name{fact2woe}
\alias{factor2woe}
\title{Factor to Weight of Evidence}
\description{Transform factorise data to weight of evidence}
\usage{factor2woe(segm, woe) }
\arguments{\item{segm}{ INTEGER / VECTOR. Number of segments to factorise numerical
variables.}\item{woe}{ woe}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
