\name{Factorise}
\alias{Factorise}
\title{Factorise variable}
\description{Factorise numerical variables according with defined bins}

\usage{Factorise(X, nseg, 
seg.type = c("freq_equal", "width_equal"), 
na.replace = NULL) }

\arguments{\item{X}{ X}
\item{nseg}{ INTEGER / VECTOR. Number of segments to factorise numerical variables.}
\item{seg.type}{ CHARACTER. Type of segments to create. (Default = "equal frequencies")}
\item{na.replace}{ CHARACTER / NUMERIC. Value to replace missing. \cr 
If NULL missing values are not considered in the computation.}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
