\name{factor2woe}

\alias{factor2woe}

\title{Factor to Weight of Evidence}

\description{Transform factorise data to weight of evidence}

\usage{factor2woe(segm, woe) }

\arguments{\item{segm}{ }\item{woe}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
