\name{Factorise}

\alias{Factorise}

\title{Factorise variable}

\description{Factorise numerical variables according with defined bins}

\usage{Factorise(X, nseg, seg.type = c("freq_equal", "width_equal"), na.replace = NULL) }

\arguments{\item{X}{ }\item{nseg}{ }\item{seg.type}{ }\item{na.replace}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
