\name{FFT}
\alias{FFT}

\title{Custom Fast Fourier transformation}

\description{Compute Custom Fast Fourier transformation}

\usage{FFT(X, Fs = 1, half = FALSE, window = NULL, plot = TRUE, optimised = TRUE, ...) }

\arguments{ \item{X}{ }\item{Fs}{ }\item{half}{ }\item{window}{ }\item{plot}{ }\item{optimised}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
