\name{fin.plot}
\alias{fin.plot}

\title{Plot financial time series}

\description{}

\usage{fin.plot(X, top.vars = c("Close", "High", "Low"), bottom.vars = "Volume", style =
c("default", "candlestick"), snames = attr(X, "SName"), xlabels = rownames(X), main = "", main2 = "",
ytitle = "", ytitle2 = "", theme.top = getCurrentTheme(), overrides = list(...), theme.bottom =
getCurrentTheme(), overrides2 = NULL, ...) }

\arguments{ \item{X}{ }\item{top.vars}{ }\item{bottom.vars}{ }\item{style}{ }\item{snames}{
}\item{xlabels}{ }\item{main}{ }\item{main2}{ }\item{ytitle}{ }\item{ytitle2}{ }\item{theme.top}{
}\item{overrides}{ }\item{theme.bottom}{ }\item{overrides2}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
