\name{FirstHit}
\alias{FirstHit}

\title{First Hit time barrier (Brownian motion)}

\description{Calculate expected time of the First Hitting for a Brownian motion}

\usage{FirstHit(B, S0, mi, geom = FALSE, sigma = NULL) }

\arguments{ \item{B}{ }\item{S0}{ }\item{mi}{ }\item{geom}{ }\item{sigma}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
