\name{FitRes.VecAr}
\alias{FitRes.VecAr}

\title{Fiited / Residual for VAR}

\description{Get Fitted values and Residuals from a VAR model}

\usage{FitRes.VecAr(Y, Coefs, ar.lags) }

\arguments{ \item{Y}{ }\item{Coefs}{ }\item{ar.lags}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
