\name{flushLogBuffer}
\alias{flushLogBuffer}

\title{Logging utilities}

\description{Tool for Logging operations}

\usage{flushLogBuffer(console = FALSE, logfile = getLogFile(env = env), env = getOption("RAdamant")) }

\arguments{ \item{console}{ }\item{logfile}{ }\item{env}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
