\name{forcidx}
\alias{forcidx}
\title{Force index}
\description{Compute Force index (Technical Analysis)}

\usage{forcidx(X, Volume, lag = 5, sth = TRUE, 
sth.lag = 13, mov = sma, plot = FALSE, ...) }

\arguments{ \item{X}{ X}\item{Volume}{ Volume}\item{lag}{ INTEGER. Number of lag
periods.}\item{sth}{ sth}
\item{sth.lag}{ sth.lag}\item{mov}{ mov}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{
Further arguments to or from othermethods}}

\details{The Force Index measures the force of bulls behind every rally and of bears behind every
decline. The Force Index attempts to quantify the force of every price move according to its
direction, distance, and trading volume.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
