\name{formula.mreg}
\alias{formula.mreg}

\title{Extract formula from Multi-Regression (mreg) object}

\description{Extract formula from Multi-Regression (mreg) object}

\usage{formula.mreg(mod, ...) }

\arguments{ \item{mod}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
