\name{formula.reg}
\alias{formula.reg}

\title{Extract formula from regression (reg) object}

\description{Extract formula from regression (reg) object}

\usage{formula.reg(mod, ...) }

\arguments{ \item{mod}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
