\name{fresvar}

\alias{fitted.VecAr}

\title{Fiited / Residual for VAR}

\description{Get Fitted values and Residuals from a VAR model}

\usage{
\method{fitted}{VecAr}(object, Coefs, ar.lags, ...) 
}

\arguments{ \item{object}{ object}\item{Coefs}{ Coefs}\item{ar.lags}{ ar.lags}\item{...}{ Further
arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
