\name{FSE.VecAr}
\alias{FSE.VecAr}

\title{VAR Forecast Standard Error }

\description{Compute forecast standard error for VAR model}

\usage{FSE.VecAr(X, steps) }

\arguments{ \item{X}{ }\item{steps}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
