\name{fullP}
\alias{fullP}

\title{Full price}

\description{Compute Full price (Technical Analysis)}

\usage{fullP(Close, Open, High, Low, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{Open}{ }\item{High}{ }\item{Low}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
