\name{func.comment.idx}
\alias{func.comment.idx}

\title{Function comment}

\description{Given an input file, this functions created an index based commented version of the file.}

\usage{func.comment.idx(control.df = data.frame(FNAME = c(), FCODE = c(), AREA = c(), SECTION = c(),
CLASS = c()), infile = NULL, incode = NULL, outfile = NULL, max.dgt = 3) }

\arguments{ \item{control.df}{ }\item{infile}{ }\item{incode}{ }\item{outfile}{ }\item{max.dgt}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
