\name{func.line.cnt}
\alias{func.line.cnt}

\title{Function line counting}

\description{Given a package name or a list of functions, for each function X
in the package or the list it counts the lines of code, the number of subcalls
made to any other function Y of the list/package and the number of other
functions that make calls to the function X. \cr 
Results are plotted if requested.}

\usage{func.line.cnt(package = NULL, plot = TRUE, qtz.type = "NONE", qtz.nbins = 10, qtz.cutoff =
30) }

\arguments{ \item{package}{ }\item{plot}{ }\item{qtz.type}{ }\item{qtz.nbins}{ }\item{qtz.cutoff}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
