\name{Gain}

\alias{Gain}
\alias{Gain.scorecard}

\title{Gain plot}

\description{Gain plot for a classification model}

\usage{
Gain(x, ...) \cr
Gain.scorecard(sc, pc = 0.1, ...) 
}

\arguments{
\item{x}{ }
\item{sc}{ }
\item{pc}{ }
\item{...}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
