\name{Garch.default}
\alias{Garch.default}

\title{Garch – default method}

\description{Default method for Garch model }

\usage{Garch.default(x, Y = NULL, order = c(alpha = 1, beta = 1), type = c("garch", "tgarch",
"egarch"), prob = c("norm", "ged", "t"), r = 3, par.init = c("default", "res.std"), phi = 0.5) }

\arguments{ \item{x}{ }\item{Y}{ }\item{order}{ }\item{type}{ }\item{prob}{ }\item{r}{ }\item{par.init}{ }\item{phi}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
