\name{Garch.proc}
\alias{Garch.proc}

\title{Garch process simulation}

\description{Simulate a Garch process and return fitted series}

\usage{Garch.proc(pars, order, res, type = c("garch", "tgarch", "egarch"), r, prob) }

\arguments{ \item{pars}{ }\item{order}{ }\item{res}{ }\item{type}{ }\item{r}{ }\item{prob}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
