\name{Garch}
\alias{Garch}

\title{Garch – method}

\description{Generic method for Garch model class}

\usage{Garch(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
