\name{get.acf.ci}
\alias{get.acf.ci}

\title{Normal confidence intervals for correlation}

\description{Compute the Normal confidence intervals for correlation and partial autocorrelation data}

\usage{get.acf.ci(X, ci = 0.95) }

\arguments{ \item{X}{ }\item{ci}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
