\name{get.fs}
\alias{get.fs}

\title{Download Financial Series data from Yahoo!}

\description{}

\usage{get.fs(symbol = NULL, SName = NULL, from = as.Date("1950-01-01"), to = Sys.Date(),
strip.spaces = TRUE, strip.char = ".") }

\arguments{ \item{symbol}{ }\item{SName}{ }\item{from}{ }\item{to}{ }\item{strip.spaces}{ }\item{strip.char}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
