\name{get.lm.weights}
\alias{get.lm.weights}

\title{Extract weights percentages of the coefficients of a linear model}

\description{}

\usage{get.lm.weights(mod, pct = FALSE) }

\arguments{ \item{mod}{ }\item{pct}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
