\name{get.plot.layout}
\alias{get.plot.layout}

\title{Get plot layout}

\description{Plotting tools}

\usage{get.plot.layout(N = 1, theme.params = getCurrentTheme(), overrides = NULL) }

\arguments{ \item{N}{ }\item{theme.params}{ }\item{overrides}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
