\name{get.plot.params}
\alias{get.plot.params}

\title{Get plot parameters}

\description{Plotting tools}

\usage{get.plot.params(class = NULL, type = NULL, ...) }

\arguments{ \item{class}{ }\item{type}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
