\name{get.predictors}
\alias{get.predictors}

\title{Extract the column names of the regression terms of a linear model}

\description{}

\usage{get.predictors(mod) }

\arguments{ \item{mod}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
