\name{getCurrentTheme}
\alias{getCurrentTheme}

\title{Graphics themes}

\description{Plotting tools}

\usage{getCurrentTheme(env = getOption("RAdamant")) }

\arguments{ \item{env}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
