\name{getlmwgh}
\alias{get.lm.weights}
\title{Extract weights percentages of the coefficients of a linear model}
\description{Extract weights percentages of the coefficients of a linear model}
\usage{get.lm.weights(mod, pct = FALSE) }
\arguments{ \item{mod}{ mod}\item{pct}{ pct}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


