\name{getLogWarning}
\alias{getLogWarning}

\title{Logging utilities}

\description{Tool for Logging operations}

\usage{getLogWarning(env = getOption("RAdamant")) }

\arguments{ \item{env}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
