\name{getpred}
\alias{get.predictors}
\title{Extract the column names of the regression terms of a linear model}
\description{Extract the column names of the regression terms of a linear model}
\usage{get.predictors(mod) }
\arguments{\item{mod}{ mod}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


