\name{getProjectionMatrix}
\alias{getProjectionMatrix}

\title{Return 3D Projection matrix}

\description{Plotting tools}

\usage{getProjectionMatrix(env = getOption("RAdamant")) }

\arguments{ \item{env}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
