\name{gev.ci}
\alias{gev.ci}

\title{GEV - Distribution fitting and Confidence Intervals}

\description{}

\usage{gev.ci(Xbmax, mu = 0, xi = 0.1, sigma = 1, alpha = 0.01, df = 3, ...) }

\arguments{ \item{Xbmax}{ }\item{mu}{ }\item{xi}{ }\item{sigma}{ }\item{alpha}{ }\item{df}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
