\name{gev.contour}
\alias{gev.contour}

\title{GEV - Joint Confidence Intervals by Profile Likelihood}

\description{}

\usage{gev.contour(Xbmax, mu = 0, xi = 0.1, sigma = 1, alpha = 0.01, df = 3, ...) }

\arguments{ \item{Xbmax}{ }\item{mu}{ }\item{xi}{ }\item{sigma}{ }\item{alpha}{ }\item{df}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
