\name{gev.like}
\alias{gev.like}

\title{GEV - Log Likelihood}

\description{}

\usage{gev.like(parms, Xbmax, ...) }

\arguments{ \item{parms}{ }\item{Xbmax}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
