\name{gev.ml}
\alias{gev.ml}

\title{GEV - Maximum Likelihood Parameters Estimation}

\description{}

\usage{gev.ml(Xbmax, init = c(0, 0.1, 1), ...) }

\arguments{ \item{Xbmax}{ }\item{init}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
