\name{gev.mu.constraint}
\alias{gev.mu.constraint}

\title{GEV - Domain range for the mu parameter}

\description{}

\usage{gev.mu.constraint(parms, type = c("left", "right", "both"), Xbmax, ...) }

\arguments{ \item{parms}{ }\item{type}{ }\item{Xbmax}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
