\name{gev.sigma.constraint}
\alias{gev.sigma.constraint}

\title{GEV - Domain range for the sigma parameter}

\description{}

\usage{gev.sigma.constraint(parms, type = c("left", "right", "both"), Xbmax, parm.type = c("mu", 
"VaR", "ES"), prob = 0.01, ...) }

\arguments{ \item{parms}{ }\item{type}{ }\item{Xbmax}{ }\item{parm.type}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
