\name{gev.VaR.ci}
\alias{gev.VaR.ci}

\title{GEV - VaR calculation and Confidence Intervals}

\description{}

\usage{gev.VaR.ci(Xbmax, VaR = sum(gev.VaR.constraint(parms = c(0, xi, sigma), type = "both", Xbmax
= Xbmax, prob = prob))/2, xi = 0.1, sigma = 1, alpha = 0.01, df = 3, prob = alpha[1], ...) }

\arguments{ \item{Xbmax}{ }\item{VaR}{ }\item{xi}{ }\item{sigma}{ }\item{alpha}{ }\item{df}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
