\name{gev.VaR.constraint}
\alias{gev.VaR.constraint}

\title{GEV - Domain range for the VaR parameter}

\description{}

\usage{gev.VaR.constraint(parms, type = c("left", "right", "both"), Xbmax, prob = 0.01, ...) }

\arguments{ \item{parms}{ }\item{type}{ }\item{Xbmax}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
