\name{gev.VaR.like}
\alias{gev.VaR.like}

\title{GEV - VaR Log Likelihood}

\description{}

\usage{gev.VaR.like(parms, Xbmax, prob = 0.01, ...) }

\arguments{ \item{parms}{ }\item{Xbmax}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
