\name{gev.VaR}
\alias{gev.VaR}

\title{GEV - VaR calculation}

\description{}

\usage{gev.VaR(Xbmax, mu = NULL, xi = NULL, sigma = NULL, prob = 0.01, ...) }

\arguments{ \item{Xbmax}{ }\item{mu}{ }\item{xi}{ }\item{sigma}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
