\name{gev.xi.constraint}
\alias{gev.xi.constraint}

\title{GEV - Domain range for the xi parameter}

\description{}

\usage{gev.xi.constraint(parms, type = c("left", "right", "both"), Xbmax, parm.type = c("mu", "VaR",
"ES"), prob = 0.01) }

\arguments{ \item{parms}{ }\item{type}{ }\item{Xbmax}{ }\item{parm.type}{ }\item{prob}{ }\item{...}{
}}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
