\name{gevar}
\alias{gev.VaR}
\title{GEV - VaR calculation}
\description{GEV - VaR calculation}
\usage{gev.VaR(Xbmax, mu = NULL, xi = NULL, sigma = NULL, prob = 0.01, ...) }
\arguments{\item{Xbmax}{ Xbmax}\item{mu}{ mu}\item{xi}{ xi}\item{sigma}{ sigma}\item{prob}{
prob}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


