\name{gevarcst}
\alias{gev.VaR.constraint}
\title{GEV - Domain range for the VaR parameter}
\description{GEV - Domain range for the VaR parameter}
\usage{gev.VaR.constraint(parms, type = c("left", "right", "both"), 
Xbmax, prob = 0.01, ...) }
\arguments{
\item{parms}{ parms}
\item{type}{ type}
\item{Xbmax}{ Xbmax}
\item{prob}{ prob}
\item{...}{
Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


