\name{gevark}
\alias{gev.VaR.like}
\title{GEV - VaR Log Likelihood}
\description{GEV - VaR Log Likelihood}
\usage{gev.VaR.like(parms, Xbmax, prob = 0.01, ...) }
\arguments{\item{parms}{ parms}\item{Xbmax}{ Xbmax}\item{prob}{ prob}\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


