\name{gevci}
\alias{gev.ci}
\title{GEV - Distribution fitting and Confidence Intervals}
\description{GEV - Distribution fitting and Confidence Intervals}
\usage{gev.ci(Xbmax, mu = 0, xi = 0.1, sigma = 1, alpha = 0.01, df = 3, ...) }
\arguments{\item{Xbmax}{ Xbmax}\item{mu}{ mu}\item{xi}{ xi}\item{sigma}{ sigma}\item{alpha}{
alpha}\item{df}{ df}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


