\name{gevcont}
\alias{gev.contour}
\title{GEV - Joint Confidence Intervals by Profile Likelihood}
\description{GEV - Joint Confidence Intervals by Profile Likelihood}
\usage{gev.contour(Xbmax, mu = 0, xi = 0.1, sigma = 1, alpha = 0.01, df = 3, ...) }
\arguments{\item{Xbmax}{ Xbmax}\item{mu}{ mu}\item{xi}{ xi}\item{sigma}{ sigma}\item{alpha}{
alpha}\item{df}{ df}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
