\name{gevmcst}
\alias{gev.mu.constraint}
\title{GEV - Domain range for the mu parameter}
\description{GEV - Domain range for the mu parameter}

\usage{gev.mu.constraint(parms
                                , type = c("left", "right", "both")
                                , Xbmax
                                , ...) 
                                }

\arguments{
\item{parms}{ parms}
\item{type}{ type}
\item{Xbmax}{ Xbmax}
\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


