\name{gevml}
\alias{gev.ml}
\title{GEV - Maximum Likelihood Parameters Estimation}
\description{GEV - Maximum Likelihood Parameters Estimation}
\usage{gev.ml(Xbmax, init = c(0, 0.1, 1), ...) }
\arguments{
\item{Xbmax}{ Xbmax}
\item{init}{ init}
\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


