\name{gevsicst}
\alias{gev.sigma.constraint}
\title{GEV - Domain range for the sigma parameter}

\description{GEV - Domain range for the sigma parameter}

\usage{gev.sigma.constraint(parms
                                    , type = c("left", "right", "both")
                                    , Xbmax
                                    , parm.type = c("mu", "VaR", "ES")
                                    , prob = 0.01
                                    , ...) 
                                    }

\arguments{
\item{parms}{ parms}
\item{type}{ type}
\item{Xbmax}{ Xbmax}
\item{parm.type}{
parm.type}
\item{prob}{ prob}
\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


