\name{GKgamma}
\alias{GKgamma}
\title{GK Gamma}
\description{GK Gamma statistic}
\usage{GKgamma(target, pred, ...) }
\arguments{\item{target}{ target}\item{pred}{ pred}\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
