\name{gmean}
\alias{gmean}

\title{Geometric mean}

\description{Compute the geometric mean for each column of X}

\usage{gmean(X, ...) }

\arguments{ \item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
