\name{gpd.contour}
\alias{gpd.contour}

\title{GPD - Joint Confidence Intervals by Profile Likelihood}

\description{}

\usage{gpd.contour(Xtail, trsh = 0, xi = 0.1, sigma = 1, alpha = 0.01, df = 2, ...) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{xi}{ }\item{sigma}{ }\item{alpha}{ }\item{df}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
