\name{gpd.ES.ci}
\alias{gpd.ES.ci}

\title{GPD - ES calculation and Confidence Intervals}

\description{}

\usage{gpd.ES.ci(Xtail, trsh = 0, ES = trsh + 10^-5, xi = 0.1, alpha = 0.01, df = 2, N, prob =
alpha[1], ...) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{ES}{ }\item{xi}{ }\item{alpha}{ }\item{df}{ }\item{N}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
