\name{gpd.ES.ml}
\alias{gpd.ES.ml}

\title{GPD - Maximum Likelihood ES Estimation}

\description{}

\usage{gpd.ES.ml(Xtail, trsh = 0, N, init = c(1, 0.1), ...) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{N}{ }\item{init}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
