\name{gpd.ES.range}
\alias{gpd.ES.range}
\title{GPD - ES range grid for contour calculation}
\description{GPD - ES range grid for contour calculation}
\usage{gpd.ES.range(Xtail, trsh = 0, ES = trsh + 10^-5, 
xi = 0.1, alpha = 0.01, df = 2, N, prob = alpha[1], ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{ES}{ ES}\item{xi}{ xi}\item{alpha}{
alpha}\item{df}{ df}\item{N}{ N}\item{prob}{ prob}\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
