\name{gpd.ES}
\alias{gpd.ES}

\title{GPD - Expected Shortfall (ES) calculation}

\description{}

\usage{gpd.ES(Xtail, trsh = 0, xi = NULL, sigma = NULL, N, prob = 0.01, ...) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{xi}{ }\item{sigma}{ }\item{N}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
