\name{gpd.like}
\alias{gpd.like}

\title{GPD - Log Likelihood}

\description{}

\usage{gpd.like(parms, Xtail, trsh = 0, ...) }

\arguments{ \item{parms}{ }\item{Xtail}{ }\item{trsh}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
