\name{gpd.ml}
\alias{gpd.ml}

\title{GPD - Maximum Likelihood Parameters Estimation}

\description{}

\usage{gpd.ml(Xtail, trsh = 0, init = c(0.1, 1), ...) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{init}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
