\name{gpd.sigma.constraint}
\alias{gpd.sigma.constraint}

\title{GPD - Domain range for the sigma parameter}

\description{}

\usage{gpd.sigma.constraint(parms, type = c("left", "right", "both"), Xtail, trsh = 0, ...) }

\arguments{ \item{parms}{ }\item{type}{ }\item{Xtail}{ }\item{trsh}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
