\name{gpd.surface}
\alias{gpd.surface}

\title{GPD - Log Likelihood 3D surface}

\description{}

\usage{gpd.surface(xi = NULL, sigma = NULL, Xtail, trsh = 0, grid.size = 100, alpha = 0.01, ...) }

\arguments{ \item{xi}{ }\item{sigma}{ }\item{Xtail}{ }\item{trsh}{ }\item{grid.size}{ }\item{alpha}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
