\name{gpd.VaR.like}
\alias{gpd.VaR.like}

\title{GPD - VaR Log Likelihood}

\description{}

\usage{gpd.VaR.like(parms, Xtail, trsh = 0, N, prob = 0.01, ...) }

\arguments{ \item{parms}{ }\item{Xtail}{ }\item{trsh}{ }\item{N}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
