\name{gpd.VaR.surface}
\alias{gpd.VaR.surface}

\title{GPD - Log Likelihood 3D surface as a function of VaR}

\description{}

\usage{gpd.VaR.surface(VaR = NULL, xi = NULL, Xtail, trsh = 0, N, prob = 0.01, grid.size = 100,
alpha = 0.01, ...) }

\arguments{ \item{VaR}{ }\item{xi}{ }\item{Xtail}{ }\item{trsh}{ }\item{N}{ }\item{prob}{
}\item{grid.size}{ }\item{alpha}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
