\name{gpd.xi.constraint}
\alias{gpd.xi.constraint}

\title{GPD - Domain range for the xi parameter}

\description{}

\usage{gpd.xi.constraint(parms, type = c("left", "right", "both"), Xtail, trsh = 0, parm.type =
c("sigma", "VaR", "ES"), prob = 0.01, ...) }

\arguments{ \item{parms}{ }\item{type}{ }\item{Xtail}{ }\item{trsh}{ }\item{parm.type}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
