\name{gpdboot}
\alias{gpdboot}

\title{GPD - parameters bootstrapping}

\description{}

\usage{gpdboot(Xtail, trsh = 0, xi = NULL, sigma = NULL, nboots = 100, ...) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{xi}{ }\item{sigma}{ }\item{nboots}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
