\name{gpdci}
\alias{gpd.ci}
\title{GPD - Distribution fitting and Confidence Intervals}
\description{GPD - Distribution fitting and Confidence Intervals}
\usage{gpd.ci(Xtail, trsh = 0, xi = 0.1, sigma = 1, alpha = 0.01, df = 2, ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{xi}{ xi}\item{sigma}{ sigma}\item{alpha}{
alpha}\item{df}{ df}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


