\name{gpdes}
\alias{gpd.ES}
\title{GPD - Expected Shortfall (ES) calculation}
\description{GPD - Expected Shortfall (ES) calculation}
\usage{gpd.ES(Xtail, trsh = 0, xi = NULL, sigma = NULL, N, prob = 0.01, ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{xi}{ xi}\item{sigma}{ sigma}\item{N}{
N}\item{prob}{ prob}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


