\name{gpdesci}
\alias{gpd.ES.ci}
\title{GPD - ES calculation and Confidence Intervals}
\description{GPD - ES calculation and Confidence Intervals}
\usage{gpd.ES.ci(Xtail, trsh = 0, ES = trsh + 10^-5, xi = 0.1, 
alpha = 0.01, df = 2, N, prob = alpha[1], ...) }
\arguments{
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{ES}{ ES}
\item{xi}{ xi}
\item{alpha}{
alpha}
\item{df}{ df}
\item{N}{ N}
\item{prob}{ prob}
\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


