\name{gpdescnt}
\alias{gpd.ES.contour}
\title{GPD - ES Joint Confidence Intervals by Profile Likelihood}
\description{GPD - ES Joint Confidence Intervals by Profile Likelihood}

\usage{gpd.ES.contour(Xtail, trsh = 0, ES = trsh + 10^-5, 
xi = 0.1, alpha = 0.01, df = 2, N, prob = alpha[1], ...) }
\arguments{
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{ES}{ ES}
\item{xi}{ xi}
\item{alpha}{
alpha}
\item{df}{ df}
\item{N}{ N}
\item{prob}{ prob}
\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


