\name{gpdescst}
\alias{gpd.ES.constraint}
\title{GPD - Domain range for the ES parameter}
\description{GPD - Domain range for the ES parameter}
\usage{gpd.ES.constraint(parms, type = c("left", "right", "both"), trsh = 0, ...) }
\arguments{\item{parms}{ parms}\item{type}{ type}\item{trsh}{ trsh}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
