\name{gpdesfce}
\alias{gpd.ES.surface}
\title{GPD - Log Likelihood 3D surface as a function of Expected Shortfall}
\description{GPD - Log Likelihood 3D surface as a function of Expected Shortfall}
\usage{gpd.ES.surface(ES = NULL, xi = NULL, Xtail, 
trsh = 0, N, prob = 0.01, grid.size = 100, alpha = 0.01, ...) }
\arguments{\item{ES}{ ES}\item{xi}{ xi}\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{N}{
N}\item{prob}{ prob}\item{grid.size}{ grid.size}\item{alpha}{ alpha}\item{...}{ Further arguments to
or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
