\name{gpd.ES.like}
\alias{gpd.ES.like}
\title{GPD - ES Log Likelihood}
\description{GPD - ES Log Likelihood}
\usage{gpd.ES.like(parms, Xtail, trsh = 0, N, prob = 0.01, ...) }
\arguments{\item{parms}{ parms}\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{N}{ N}\item{prob}{
prob}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
