\name{gpd.ES.ml}
\alias{gpd.ES.ml}
\title{GPD - Maximum Likelihood ES Estimation}
\description{GPD - Maximum Likelihood ES Estimation}
\usage{gpd.ES.ml(Xtail, trsh = 0, N, init = c(1, 0.1), ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{N}{ N}\item{init}{ init}\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
