\name{gpdlk}
\alias{gpd.like}
\title{GPD - Log Likelihood}
\description{GPD - Log Likelihood}
\usage{gpd.like(parms, Xtail, trsh = 0, ...) }
\arguments{
\item{parms}{ parms}
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


