\name{gpdml}
\alias{gpd.ml}
\title{GPD - Maximum Likelihood Parameters Estimation}
\description{GPD - Maximum Likelihood Parameters Estimation}
\usage{gpd.ml(Xtail, trsh = 0, init = c(0.1, 1), ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{init}{ init}\item{...}{ Further arguments to
or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
