\name{gpdrng}
\alias{gpd.range}
\title{GPD - Parameters range grid for contour calculation}

\description{GPD - Parameters range grid for contour calculation}

\usage{gpd.range(Xtail
                    , trsh = 0
                    , xi = 0.1
                    , sigma = 1
                    , alpha = 0.01
                    , df = 2
                    , ...) 
                    }

\arguments{
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{xi}{ xi}
\item{sigma}{ sigma}
\item{alpha}{
alpha}
\item{df}{ df}
\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


