\name{gpdsfc}
\alias{gpd.surface}
\title{GPD - Log Likelihood 3D surface}
\description{GPD - Log Likelihood 3D surface}
\usage{gpd.surface(xi = NULL, sigma = NULL, Xtail, 
trsh = 0, grid.size = 100, alpha = 0.01, ...) }
\arguments{
\item{xi}{ xi}
\item{sigma}{ sigma}
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{grid.size}{
grid.size}
\item{alpha}{ alpha}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


