\name{gpd.VaR}
\alias{gpd.VaR}
\title{GPD - VaR calculation}
\description{GPD - VaR calculation}
\usage{gpd.VaR(Xtail, trsh = 0, xi = NULL, sigma = NULL, N, prob = 0.01, ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{xi}{ xi}\item{sigma}{ sigma}\item{N}{
N}\item{prob}{ prob}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
