\name{gpdvarci}
\alias{gpd.VaR.ci}
\title{GPD - VaR calculation and Confidence Intervals}
\description{GPD - VaR calculation and Confidence Intervals}
\usage{gpd.VaR.ci(Xtail, trsh = 0, VaR = trsh + 10^-5, 
xi = 0.1, alpha = 0.01, df = 2, N, prob = alpha[1], ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{VaR}{ VaR}\item{xi}{ xi}\item{alpha}{
alpha}\item{df}{ df}\item{N}{ N}\item{prob}{ prob}\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


