\name{gpdvarct}
\alias{gpd.VaR.constraint}
\title{GPD - Domain range for the VaR parameter}
\description{GPD - Domain range for the VaR parameter}
\usage{gpd.VaR.constraint(parms, type = c("left", "right", "both"), trsh = 0, ...) }
\arguments{\item{parms}{ parms}\item{type}{ type}\item{trsh}{ trsh}\item{...}{ Further arguments to
or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
