\name{gpdvarml}
\alias{gpd.VaR.ml}
\title{GPD - Maximum Likelihood VaR Estimation}
\description{GPD - Maximum Likelihood VaR Estimation}
\usage{gpd.VaR.ml(Xtail, trsh = 0, N, init = c(1, 0.1), ...) }
\arguments{
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{N}{ N}
\item{init}{ init}
\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


