\name{gpdxicst}
\alias{gpd.xi.constraint}
\title{GPD - Domain range for the xi parameter}
\description{GPD - Domain range for the xi parameter}

\usage{gpd.xi.constraint(parms, type = c("left", "right", "both"), 
Xtail, trsh = 0, N, parm.type = c("sigma", "VaR", "ES"), 
prob = 0.01, ...) }

\arguments{
\item{parms}{ parms}
\item{type}{ type}
\item{Xtail}{ Xtail}
\item{trsh}{ trsh}
\item{N}{
N}
\item{parm.type}{ parm.type}
\item{prob}{ prob}
\item{...}{ Further arguments to or from other
methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


