\name{grad}
\alias{grad}

\title{Compute numerical gradient of a function}

\description{Plotting tools}

\usage{grad(func = NULL, x, scalar = TRUE, eps = sqrt(.Machine$double.neg.eps), ...) }

\arguments{ \item{func}{ }\item{x}{ }\item{scalar}{ }\item{eps}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
