\name{GrangCas}
\alias{GrangCas}

\title{Granger Causality test}

\description{Perform Granger causality test for parameters of VAR model}

\usage{GrangCas(X, cause = NULL) }

\arguments{ \item{X}{ }\item{cause}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
