\name{grautil}
\alias{transition}
\alias{gradient}
\alias{jet.colors}
\alias{override.list}
\alias{set.bg}
\alias{set.bg3d}
\alias{create.empty.plot}
\alias{optimize.polycords}
\alias{shade.plot} 
\alias{comma.Fmt} 
\alias{comma.kFmt} 
\alias{comma.mFmt}
\alias{SI.format} 
\alias{apply.format} 

\title{RAdamant Plot Utility Functions}
\description{Utility functions for internal plotting functions.}

\author{RAdamant Development Team \email{team@r-adamant.org}}


