\name{he_as}
\alias{he_as}

\title{Heikin – Ashi techniques}

\description{Compute Heikin – Ashi techniques (Technical Analysis)}

\usage{he_as(Close, Open, High, Low, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{Open}{ }\item{High}{ }\item{Low}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
