\name{heas}
\alias{he_as}
\title{Heikin - Ashi techniques}
\description{Compute Heikin - Ashi techniques (Technical Analysis)}
\usage{he_as(Close, Open, High, Low, plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{Open}{Vector. Open price.}
\item{High}{Vector.
High price.}
\item{Low}{Vector. Low price.}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}



\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


