\name{hhv}
\alias{hhv}
\title{Highest high}
\description{Compute Highest high (Technical Analysis)}
\usage{hhv(X, lag, na.rm = TRUE) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{na.rm}{ na.rm}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


