\name{hroi}
\alias{hroi}

\title{Historical Returns on Investments}

\description{}

\usage{hroi(X, lag = 1, mode = c("auto", "range", "selected"), autolag.start = 1, range.step = 1,
log = TRUE, VaR.type = "norm", p = 0.05, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{mode}{ }\item{autolag.start}{ }\item{range.step}{ }\item{log}{ }\item{VaR.type}{ }\item{p}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
