\name{Ichkh}
\alias{Ichkh}

\title{Ichimoku Kinko Hyo}

\description{Compute Ichimoku Kinko Hyo (Technical Analysis)}

\usage{Ichkh(Close, High, Low, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
