\name{inertia}
\alias{Inertia}
\title{Inertia oscillator}
\description{Compute Inertia oscillator (Technical Analysis)}
\usage{Inertia(X, lag, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{...}{ Further arguments to
or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
