\name{input2woe}

\alias{input2woe}

\title{Data to Weight of Evidence}

\description{Transform input data according to weight of evidence}

\usage{input2woe(data, nseg, woe, na.replace, seg.type = c("freq_equal", "width_equal")) }

\arguments{\item{data}{ }\item{nseg}{ }\item{woe}{ }\item{na.replace}{ }\item{seg.type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org} }

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
