\name{inv.logit}
\alias{inv.logit}

\title{Inverse Logit transformation}

\description{Inverse Logit transformation}

\usage{inv.logit(y) }

\arguments{ \item{y}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
