\name{InvPP}
\alias{InvPP}

\title{Peizer-Pratt Inversion formula}

\description{Peizer-Pratt Inversion formula}

\usage{InvPP(z, n) }

\arguments{ \item{z}{ }\item{n}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
