\name{IRS.VecAr}
\alias{IRS.VecAr}

\title{VAR Impulse response}

\description{Compute and plot Impulse response function calculated for VAR model}

\usage{IRS.VecAr(X, imp, resp = NULL, steps = 5, cum = TRUE, ortho = FALSE) }

\arguments{ \item{X}{ }\item{imp}{ }\item{resp}{ }\item{steps}{ }\item{cum}{ }\item{ortho}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
