\name{is.fs}
\alias{is.fs}

\title{Check for inheritance from Financial Series class}

\description{}

\usage{is.fs(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
