\name{is.fs}
\alias{is.fs}
\title{Check for inheritance from Financial Series class}
\description{Check for inheritance from Financial Series class}
\usage{is.fs(X) }
\arguments{\item{X}{ X}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
