\name{jrbtree}
\alias{JR.BinTree}
\title{JR Binomial Tree}
\description{Option evaluation with Jarrow and Rudd Binomial Tree}

\usage{JR.BinTree(Nsteps, p, under, strike, rfr, sigma, 
maty, yield, life, ret.steps = FALSE) }

\arguments{\item{Nsteps}{ Nsteps}\item{p}{ p}\item{under}{ under}\item{strike}{ strike}\item{rfr}{
rfr}\item{sigma}{ sigma}\item{maty}{ maty}\item{yield}{ yield}\item{life}{ life}\item{ret.steps}{
ret.steps}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
