\name{kelt}
\alias{kelt}
\title{Keltner channel}
\description{Compute Keltner channel (Technical Analysis)}
\usage{kelt(Close, High, Low, mult = 2, plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{mult}{ mult}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


