\name{KendallTau}

\alias{KendallTau}

\title{KendallTau}

\description{Calculate KendallTau statistic}

\usage{KendallTau(target, pred, ...) }

\arguments{\item{target}{ }\item{pred}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
