\name{kri}
\alias{kri}
\title{Kairi Relative Index}
\description{Compute Kairi Relative Index (Technical Analysis)}
\usage{kri(X, lag1 = 10, lag2 = 20, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag1}{ lag1}\item{lag2}{ lag2}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


