\name{Kurt_Skew}
\alias{kurt}
\alias{skew}

\title{Kurtosis and Skewness}

\description{
\code{kurt}: Compute the excess kurtosis for each column of X \cr
\code{skew}: Compute the skewness for each column of X
}

\usage{
kurt(X, pval = FALSE) \cr
skew(X, pval = FALSE)
}

\arguments{ 
\item{X}{ Matrix of numeric data series (one column per variable).}
\item{pval}{ LOGICAL. Return P-Value.}
}

\details{}

\value{Matrix of Excess Kurtosis / Skewness and P-Value}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{ \code{\link{JB.test}} }

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
