\name{kvo}
\alias{kvo}

\title{Klinger oscillator}

\description{Compute Klinger oscillator (Technical Analysis)}

\usage{kvo(Close, High = NULL, Low = NULL, Vol = NULL, cumulative = FALSE, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{Vol}{ }\item{cumulative}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
