\name{like.egarch}
\alias{like.egarch}

\title{EGARCH likelihood function}

\description{Calculate EGARCH likelihood function}

\usage{like.egarch(theta, ee, x, Y, order = c(alpha = 1, beta = 1), prob = c("norm", "ged", "t"), r) }

\arguments{ \item{theta}{ }\item{ee}{ }\item{x}{ }\item{Y}{ }\item{order}{ }\item{prob}{ }\item{r}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
