\name{like.garch}
\alias{like.garch}

\title{GARCH likelihood function}

\description{Calculate GARCH likelihood function}

\usage{like.garch(theta, ee, x, Y, order, prob = c("norm", "ged", "t"), r) }

\arguments{ \item{theta}{ }\item{ee}{ }\item{x}{ }\item{Y}{ }\item{order}{ }\item{prob}{ }\item{r}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
