\name{like.tgarch}
\alias{like.tgarch}

\title{TGARCH likelihood function}

\description{Calculate TGARCH likelihood function}

\usage{like.tgarch(theta, ee, x, Y, order, prob = c("norm","ged", "t"), r) }

\arguments{ \item{theta}{ }\item{ee}{ }\item{x}{ }\item{Y}{ }\item{order}{ }\item{prob}{ }\item{r}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
