\name{ljbgarch}

\alias{LjungBox}

\title{Ljung-Box test}

\description{Perform Ljung-Box test for residual correlation}


\usage{
LjungBox(x, lags, plot.acf = FALSE)
}

\arguments{
\item{x}{ Residual series or object of class "Garch"}
\item{lags}{ Number of lags to calculate the autocorrelation function}
\item{plot.acf}{ LOGICAL. Plot ACF.}
}



