\name{LjungBox.Garch}
\alias{LjungBox.Garch}

\title{Ljung-Box test}

\description{Perform Ljung-Box test for residual correlation}

\usage{
LjungBox.Garch(x, lags, plot.acf = FALSE)
}

\arguments{
\item{x}{ Residual series}
\item{lags}{ Number of lags to calculate the autocorrelation function}
\item{plot.acf}{ LOGICAL. Plot ACF.}
}

\details{}

\value{}

\references{}

\author{}

\note{}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
