\name{lkegarch}
\alias{like.egarch}
\title{EGARCH likelihood function}
\description{Calculate EGARCH likelihood function}
\usage{like.egarch(theta, ee, x, Y, 
order = c(alpha = 1, beta = 1), prob = c("norm", "ged", "t")) }
\arguments{\item{theta}{ theta}\item{ee}{ ee}\item{x}{ x}\item{Y}{ Y}\item{order}{
order}\item{prob}{ prob}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


