\name{lkgarch}
\alias{like.garch}
\title{GARCH likelihood function}
\description{Calculate GARCH likelihood function}
\usage{like.garch(theta, ee, x, Y, order, prob = c("norm", "ged", "t"), r) }
\arguments{\item{theta}{ theta}\item{ee}{ ee}\item{x}{ x}\item{Y}{ Y}\item{order}{
order}\item{prob}{ prob}\item{r}{ r}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
