\name{lktgarch}
\alias{like.tgarch}
\title{TGARCH likelihood function}
\description{Calculate TGARCH likelihood function}
\usage{like.tgarch(theta, ee, x, Y, order, prob = c("norm","ged", "t")) }
\arguments{\item{theta}{ theta}\item{ee}{ ee}\item{x}{ x}\item{Y}{ Y}\item{order}{
order}\item{prob}{ prob}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
