\name{llv}
\alias{llv}

\title{Lowest low}

\description{Compute Lowest low (Technical Analysis)}

\usage{llv(X, lag, na.rm = TRUE) }

\arguments{ \item{X}{ }\item{lag}{ }\item{na.rm}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
