\name{Logging}
\alias{setConsoleLogging}
\alias{setDebugLevel}
\alias{setDebugTraceLevel}
\alias{setLogBufferSize}
\alias{setLogFile}
\alias{setLogWarning}
\alias{updateLogCounter}

\title{ Logging utilities}

\description{Set parameters for the logging functionalities}

\usage{ 
setConsoleLogging(consoleLogging = TRUE, env = getOption("RAdamant")) \cr
setDebugLevel(level = 1, env = getOption("RAdamant")) \cr
setDebugTraceLevel(level = 1, env = getOption("RAdamant")) \cr
setLogBufferSize(size = 10000, env = getOption("RAdamant"), ...) \cr
setLogFile(logfile = NULL, env = getOption("RAdamant")) \cr
setLogWarning(showWarning = TRUE, env = getOption("RAdamant")) \cr
updateLogCounter(env = getOption("RAdamant"), ...) 
}

\arguments{
\item{consoleLogging}{ }
\item{env}{ }
\item{level}{ }
\item{size}{ }
\item{logfile}{ }
\item{showWarning}{ }
\item{...}{ }
}

\details{ }

\value{ }

\references{ }

\author{ RAdamant Development Team \email{team@r-adamant.org} }

\note{ }

\seealso{ }

\examples{ }

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
