\name{logit}
\alias{logit}

\title{Logit transformation}

\description{Logit transformation}

\usage{logit(x, adjust = 5e-05) }

\arguments{ \item{x}{ }\item{adjust}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
