\name{LR.BinTree}
\alias{LR.BinTree}

\title{LR Binomial Tree}

\description{Option evaluation with Leinsen and Reimer Binomial Tree}

\usage{LR.BinTree(Nsteps, under, strike, rfr, sigma, maty, yield, life, ret.steps = FALSE) }

\arguments{ \item{Nsteps}{ }\item{under}{ }\item{strike}{ }\item{rfr}{ }\item{sigma}{ }\item{maty}{ }\item{yield}{ }\item{life}{ }\item{ret.steps}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
