\name{lrbtree}
\alias{LR.BinTree}
\title{LR Binomial Tree}
\description{Option evaluation with Leinsen and Reimer Binomial Tree}
\usage{LR.BinTree(Nsteps, under, strike, rfr, 
sigma, maty, yield, life, ret.steps = FALSE) }

\arguments{ 
\item{Nsteps}{ Nsteps}
\item{under}{ under}\item{strike}{ strike}\item{rfr}{ rfr}\item{sigma}{ sigma}\item{maty}{ maty}
\item{yield}{ yield}
\item{life}{ life}
\item{ret.steps}{ ret.steps}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


