\name{macd}
\alias{macd}

\title{Moving Average Convergence / Divergence}

\description{Compute Moving Average Convergence / Divergence (Technical Analysis)}

\usage{macd(X, fast.lag = 12, slow.lag = 26, signal.lag = 14, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{fast.lag}{ }\item{slow.lag}{ }\item{signal.lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
