\name{masscum}
\alias{mass.cum}
\title{Mass indicator cumulative}
\description{Compute Mass indicator cumulative (Technical Analysis)}
\usage{mass.cum(High, Low, Close = NULL, lag = 9, plot = FALSE, ...) }
\arguments{
\item{High}{Vector. High price.}
\item{Low}{Vector. Low price.}
\item{Close}{Vector.
Close price.}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


