\name{mcgind}
\alias{mcgind}
\title{ McGinley Dynamic Indicator}
\description{Compute  McGinley Dynamic Indicator (Technical Analysis)}
\usage{mcgind(X, lag = 12, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


