\name{mcosc}
\alias{mcosc}

\title{McClellan Oscillator}

\description{Compute McClellan Oscillator (Technical Analysis)}

\usage{mcosc(X, fast.lag = 19, slow.lag = 39, hist.lag = 9, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{fast.lag}{ }\item{slow.lag}{ }\item{hist.lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
