\name{mcplot}
\alias{mcplot}

\title{Muliple correlation plot}

\description{Multiple correlation plot}

\usage{mcplot(X, hist.nclass = 10, theme.params = getCurrentTheme(), coLin = TRUE, main =
ifelse(coLin, "Co-Linearity Analysis", "Multi-Correlation Analysis"), new.device = FALSE, ...) }

\arguments{ \item{X}{ }\item{hist.nclass}{ }\item{theme.params}{ }\item{coLin}{ }\item{main}{ }\item{new.device}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
