\name{mcsi}
\alias{mcsi}

\title{McClellan Summation Index}

\description{Compute McClellan Summation Index (Technical Analysis)}

\usage{mcsi(matr, nr, nc, lag1, lag2, plot = FALSE, ...) }

\arguments{ \item{matr}{ }\item{nr}{ }\item{nc}{ }\item{lag1}{ }\item{lag2}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
