\name{Means}
\alias{gmean}
\alias{hmean}
\title{Geometric and Harmonic means}
\description{
\code{gmean}: Compute the geometric mean for each column of X \cr
\code{hmean}: Compute the harmonic mean for each column of X
}
\usage{
gmean(X, ...) 
hmean(X, ...)
}
\arguments{ 
\item{X}{ Matrix of data series (one column per variable)}
\item{...}{ Additional parameters accepted by the function sum (i.e. na.rm)}
}
\value{Matrix of harmonic / geometric means}
\author{RAdamant Development Team \email{team@r-adamant.org} }
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
