\name{Method_acf}

\alias{print.cool.acf}
\alias{plot.cool.acf}

\title{Cool.Acf functions}

\description{Method for class 'cool.acf'}

\usage{
print.cool.acf(X) 
plot.cool.acf(X) 
}

\arguments{ 
\item{X}{ Instance of class 'cool.acf'}
\item{theme.params}{ Theme parameters (DEFAULT: getCurrentTheme())}
\item{xtitle}{ Title for the x-axis (DEFAULT: "Lag")}
\item{ytitle}{ Title for the y-axis (DEFAULT: expression(rho))}
\item{overrides}{ List of parameters to override the theme. Must match by name the parameters defined by the theme (DEFAULT: NULL)}
}

\details{}

\value{Void}

\references{}

\author{ RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
