\name{Method_univar}

\alias{plot.univar}
\alias{print.univar}

\title{Methods for univariate analysis}

\description{Print and Plot methods for class 'univar'}

\usage{
plot.univar(X) 
print.univar(X) 
}

\arguments{ 
\item{X}{ Instance of class 'univar'}
\item{theme.params}{ params: Theme parameters (DEFAULT: getCurrentTheme())}
\item{overrides}{ list of parameters to override the theme. Must match by name the parameters defined by the theme (DEFAULT: NULL)}
}

\details{Void}

\value{}

\references{}

\author{ RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{ \code{\link{univar}}}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
