\name{Mflow.ind}
\alias{Mflow.ind}

\title{Money flow indicator}

\description{Compute Money flow indicator (Technical Analysis)}

\usage{Mflow.ind(Close, High, Low, Volume, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{Volume}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
