\name{mfratio}
\alias{Mflow.ratio}
\title{Money flow ratio}
\description{Compute Money flow ratio (Technical Analysis)}
\usage{Mflow.ratio(Close, High, Low, Volume, plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{Volume}{Vector. Asset traded Volume.}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


