\name{minmaxs}
\alias{Minmaxscal}
\title{Mini/Max Scale}
\description{Compute minimum / maximum scale of a vector}
\usage{Minmaxscal(x, tmin = 0, tmax = 1) }
\arguments{\item{x}{ x}\item{tmin}{ tmin}\item{tmax}{ tmax}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


