\name{mom}

\alias{mom}

\title{Momentum oscillator}

\description{Compute Momentum oscillator (Technical Analysis)}

\usage{mom(X, lag = 5, plot = TRUE, ...) }

\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}


