\name{mse}
\alias{mse}
\title{Mean squared error}
\description{Calculate Mean squared error for model evaluation }
\usage{mse(actual, predicted) }
\arguments{ \item{actual}{ actual}\item{predicted}{ predicted}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
