\name{Multi.Styles}
\alias{Multi.Styles}
\title{Multiple Style analysis (Portfolio)}
\description{Perform Style analysis for multiple time periods }
\usage{Multi.Styles(FUND, IND, W, n_clust = 5, lower = NULL, upper = NULL, ...) }
\arguments{\item{FUND}{ FUND}\item{IND}{ IND}\item{W}{ W}\item{n_clust}{ n_clust}\item{lower}{
lower}\item{upper}{ upper}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
