\name{mtfft}

\alias{plot.FFT}
\alias{print.FFT}

\title{Plot and Print FFT}

\description{Plot and Print methods for class 'FFT'}

\usage{
\method{print}{FFT}(x, ...) 

\method{plot}{FFT}(x, theme.params = getCurrentTheme(), 
overrides = list(...), shaded = TRUE,
show.periodicity = FALSE, show.legend = FALSE, 
zoom = 100, semilog = FALSE, new.device = FALSE, ...)

}

\arguments{ \item{x}{ x}\item{theme.params}{ theme.params}\item{overrides}{ overrides}\item{shaded}{
shaded}
\item{show.periodicity}{ show.periodicity}\item{show.legend}{ show.legend}\item{zoom}{ zoom}
\item{semilog}{ semilog}\item{new.device}{ new.device}\item{...}{ Further arguments to or fromother
methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
