\name{mtfs}

\alias{plot.fs}
\alias{print.fs}

\title{Methods for Financial Series class}

\description{Plot and Print methods for Financial Series class}

\usage{
\method{plot}{fs}(x, ...) 
\method{plot}{fs}(x, ...) 
}

\arguments{ \item{x}{ x}\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
