\name{mtoscil}

\alias{plot.oscil}
\alias{print.oscil}

\title{Plot function for Oscillators}

\description{Plot and Print method for Oscillators (Technical Analysis)}

\usage{
\method{print}{oscil}(x, digits = 5, ...)

\method{plot}{oscil}(x, Y = NULL, main = "", 
show.trsh = NULL, xlabels = rownames(Y), 
theme.params =getTheme(1), overrides = NULL, ...)
}

\arguments{ 
\item{x}{ x}
\item{Y}{ Y}
\item{main}{ main}
\item{show.trsh}{ show treshold}
\item{xlabels}{ xlabels}
\item{theme.params}{ them.params}
\item{overrides}{ overrides}
\item{digits}{ digits}
\item{...}{ Further arguments to or from other methods}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


