\name{Multi.Styles}
\alias{Multi.Styles}

\title{Multiple Style analysis (Portfolio)}

\description{Perform Style analysis for multiple time periods }

\usage{Multi.Styles(FUND, IND, W, n_clust = 5, lower = NULL, upper = NULL, ...) }

\arguments{ \item{FUND}{ }\item{IND}{ }\item{W}{ }\item{n_clust}{ }\item{lower}{ }\item{upper}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
