\name{newsimp.default}
\alias{newsimp.default}

\title{News impact curve }

\description{Compute News impact curve }

\usage{newsimp.default(X, theta, order, type = c("garch", "egarch", "tgarch"), plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{theta}{ }\item{order}{ }\item{type}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
