\name{newsimp.Garch}
\alias{newsimp.Garch}

\title{News impact curve – method for Garch}

\description{News Impact Curve method for Garch}

\usage{newsimp.Garch(x, plot = TRUE, ...) }

\arguments{ \item{x}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
