\name{newsimp}
\alias{newsimp}

\title{News impact curve }

\description{News Impact Curve }

\usage{newsimp(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
