\name{norm.fit}
\alias{norm.fit}

\title{Fit normal distribution}

\description{Fit normal distribution}

\usage{norm.fit(x, n = 200, range = NULL, ...) }

\arguments{ \item{x}{ }\item{n}{ }\item{range}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
