\name{norm.like}
\alias{norm.like}

\title{Normal Distribution - Log Likelihood function}

\description{}

\usage{norm.like(parms, X, ...) }

\arguments{ \item{parms}{ }\item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
