\name{normlike}
\alias{norm.like}
\title{Normal Distribution - Log Likelihood function}
\description{Normal Distribution - Log Likelihood function}
\usage{norm.like(parms, X, ...) }
\arguments{\item{parms}{ parms}\item{X}{ X}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


