\name{objgarch}

\alias{vcov.Garch}
\alias{coef.Garch}
\alias{logLik.Garch}

\title{Garch objects}

\description{Extract objects from Garch model (class "Garch")}

\usage{
\method{coef}{Garch}(object, names=TRUE, ...)
\method{logLik}{Garch}(object, ...)
\method{vcov}{Garch}(object, ...)
}

\arguments{
\item{object}{ An object of class "Garch"}
\item{names}{ Return names}
\item{...}{ Further arguments to or from other methods}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


