\name{optimize.polycords}
\alias{optimize.polycords}

\title{Graphics parameters function}

\description{Plotting tools}

\usage{optimize.polycords(x, y) }

\arguments{ \item{x}{ }\item{y}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
