\name{oscil.ARODOWN.plot.params}
\alias{oscil.ARODOWN.plot.params}

\title{Plot utils for Aroon oscillator}

\description{Compute Plot utils for Aroon oscillator (Technical Analysis)}

\usage{oscil.ARODOWN.plot.params(Osc = NULL, ...) }

\arguments{ \item{Osc}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
