\name{oscil.default}
\alias{oscil.default}

\title{Oscillator – default method}

\description{Compute Oscillator – default method (Technical Analysis)}

\usage{oscil.default(X, Y, pc = FALSE, type = "oscil") }

\arguments{ \item{X}{ }\item{Y}{ }\item{pc}{ }\item{type}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
