\name{oscil}
\alias{oscil}

\title{Oscillator – method}

\description{Compute Oscillator – method (Technical Analysis)}

\usage{oscil(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
