\name{Pchan}
\alias{Pchan}

\title{Price channel}

\description{Compute Price channel (Technical Analysis)}

\usage{Pchan(CLose, High, Low, lag = 20, na.rm = TRUE, plot = FALSE, ...) }

\arguments{ \item{CLose}{ }\item{High}{ }\item{Low}{ }\item{lag}{ }\item{na.rm}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
