\name{PDFHit}
\alias{PDFHit}

\title{Density of Hitting probability}

\description{Density for the First Hitting time}

\usage{PDFHit(t, B = 0, S0 = 0, mi, sigma, cumul = FALSE, plot = FALSE, ...) }

\arguments{ \item{t}{ }\item{B}{ }\item{S0}{ }\item{mi}{ }\item{sigma}{ }\item{cumul}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
