\name{Perf}
\alias{Perf}

\title{Performance indicator}

\description{Compute Performance indicator (Technical Analysis)}

\usage{Perf(X, ini.per = 1, cut = TRUE, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{ini.per}{ }\item{cut}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
