\name{print.Factorise}
\alias{print.Factorise}
\title{Print "factorise"}
\description{Print method for "factorise"}
\usage{
\method{print}{Factorise}(x, ...) 
}
\arguments{\item{x}{ OBJECT of class "Factorise".}\item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
