\name{print.Garch}
\alias{print.Garch}
\title{Print Garch}
\description{Print function for Garch model}
\usage{
\method{print}{Garch}(x, digits = 5, ...) 
}
\arguments{\item{x}{ x}\item{digits}{ digits}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
