\name{pgev}
\alias{pgev}

\title{Generalised Extreme Value (GEV) - Probability function}

\description{}

\usage{pgev(X, mu = 0, xi = 0.1, sigma = 1) }

\arguments{ \item{X}{ }\item{mu}{ }\item{xi}{ }\item{sigma}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
