\name{pgpd}
\alias{pgpd}

\title{Generalised Pareto Distribution (GPD) - Probability function}

\description{}

\usage{pgpd(Q, xi = 0.1, sigma = 1, trsh = 0) }

\arguments{ \item{Q}{ }\item{xi}{ }\item{sigma}{ }\item{trsh}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
