\name{print.GrangCas}
\alias{print.GrangCas}
\title{Print Granger test}
\description{Print function for Granger test}

\usage{
\method{print}{GrangCas}(x, ...) 
}

\arguments{ \item{x}{ OBJECT of class "GrangCas".}\item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
