\name{PHI.VecAr}
\alias{PHI.VecAr}

\title{VAR – PHI}

\description{Estimate PHI matrix for MA (Wold) representation of VAR model}

\usage{PHI.VecAr(X, steps, ortho = FALSE) }

\arguments{ \item{X}{ }\item{steps}{ }\item{ortho}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
