\name{plike.ci}
\alias{plike.ci}

\title{General function for profile likelihood confidence intervals calculation}

\description{}

\usage{plike.ci(ML.init = c(), flike = NULL, alpha = 0.01, df = NULL, frange = list(), par.names =
NULL, ...) }

\arguments{ \item{ML.init}{ }\item{flike}{ }\item{alpha}{ }\item{df}{ }\item{frange}{ }\item{par.names}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
