\name{plike.contour}
\alias{plike.contour}

\title{General function for profile likelihood joint confidence intervals contour}

\description{}

\usage{plike.contour(ML.init = c(), flike = NULL, alpha = 0.01, df = NULL, frange = list(),
par.names = NULL, grid.size = 100, ...) }

\arguments{ \item{ML.init}{ }\item{flike}{ }\item{alpha}{ }\item{df}{ }\item{frange}{ }\item{par.names}{ }\item{grid.size}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
