\name{plike.range}
\alias{plike.range}

\title{General range grid for contour calculation}

\description{}

\usage{plike.range(ML.init = c(), flike = NULL, alpha = 0.01, df = NULL, frange = list(), par.names
= NULL, grid.size = 100, max.iter = 100, tol = 10^-5, ...) }

\arguments{ \item{ML.init}{ }\item{flike}{ }\item{alpha}{ }\item{df}{ }\item{frange}{ }\item{par.names}{ }\item{grid.size}{ }\item{max.iter}{ }\item{tol}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
