\name{plikeci}
\alias{plike.ci}
\title{Likelihood confidence intervals calculation}
\description{General function for profile likelihood confidence intervals calculation}
\usage{
plike.ci(ML.init = c()
       , flike = NULL
       , alpha = 0.01
       , df = NULL
       , frange = list()
       , par.names = NULL
       , ...) 
}
\arguments{
\item{ ML.init}{ ML.init}
\item{ flike}{ flike}
\item{ alpha}{ alpha}
\item{ df}{df}
\item{ frange}{ frange}
\item{ par.names}{ par.names}
\item{ ...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


