\name{plikecnt}
\alias{plike.contour}
\title{Likelihood joint confidence intervals contour}
\description{General function for profile likelihood joint confidence intervals contour}

\usage{
plike.contour(ML.init = c(), flike = NULL, 
alpha = 0.01, df = NULL, frange = list(),
par.names = NULL, grid.size = 100, ...) }

\arguments{
\item{ML.init}{ ML.init}
\item{flike}{ flike}
\item{alpha}{ alpha}
\item{df}{
df}
\item{frange}{ frange}
\item{par.names}{ par.names}
\item{grid.size}{ grid.size}
\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


