\name{plikerng}
\alias{plike.range}
\title{Range grid for contour calculation}
\description{General range grid for contour calculation}

\usage{plike.range(ML.init = c(), flike = NULL, 
alpha = 0.01, df = NULL, frange = list(), par.names
= NULL, grid.size = 100, max.iter = 100, tol = 10^-5, ...) }

\arguments{\item{ML.init}{ ML.init}\item{flike}{ flike}\item{alpha}{ alpha}\item{df}{
df}\item{frange}{ frange}\item{par.names}{ par.names}\item{grid.size}{ grid.size}\item{max.iter}{
max.iter}\item{tol}{ tol}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
