\name{plot.fft}
\alias{plot.fft}

\title{Plot FFT}

\description{Plot function for class 'FFT'}

\usage{plot.fft(Xf, theme.params = getCurrentTheme(), overrides = list(...), shaded = TRUE, show.periodicity = FALSE, show.legend = FALSE, zoom = 100, semilog = FALSE, new.device = FALSE, ...) }

\arguments{ \item{Xf}{ }\item{theme.params}{ }\item{overrides}{ }\item{shaded}{ }\item{show.periodicity}{ }\item{show.legend}{ }\item{zoom}{ }\item{semilog}{ }\item{new.device}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
