\name{plot.fs}
\alias{plot.fs}

\title{Plotting function for Financial Series class}

\description{}

\usage{plot.fs(X, ...) }

\arguments{ \item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
