\name{plot.mreg}
\alias{plot.mreg}

\title{Plot function for mreg}

\description{Plot function for class 'mreg'}

\usage{plot.mreg(X, ...) }

\arguments{ \item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
