\name{plot.oscil}
\alias{plot.oscil}

\title{Plot function for Oscillators}

\description{Compute Plot function for Oscillators (Technical Analysis)}

\usage{plot.oscil(Osc, X = NULL, main = "", show.trsh = NULL, xlabels = rownames(X), theme.params =
get.theme(1), overrides = NULL, ...) }

\arguments{ \item{Osc}{ }\item{X}{ }\item{main}{ }\item{show.trsh}{ }\item{xlabels}{ }\item{theme.params}{ }\item{overrides}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
