\name{plot.reg}
\alias{plot.reg}

\title{Plot function for reg}

\description{Plot function for class 'reg'}

\usage{plot.reg(X, mode = c("response", ""), title = ifelse(X$model.type == "lm", "LS Regression",
"GLM Regression"), theme.params = getCurrentTheme(), overrides = list(...), ...) }

\arguments{ \item{X}{ }\item{mode}{ }\item{title}{ }\item{theme.params}{ }\item{overrides}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
