\name{plot.ret}
\alias{plot.ret}

\title{Plot Returns}

\description{Plot function for Returns}

\usage{plot.ret(X, style = c("line", "bar"), xlabels = rownames(X), theme.params =
getCurrentTheme(), ...) }

\arguments{ \item{X}{ }\item{style}{ }\item{xlabels}{ }\item{theme.params}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
