\name{plot.roi}
\alias{plot.roi}

\title{Plotting function for Return on Investment objects}

\description{}

\usage{plot.roi(X, main = "Historical Return on Investment", xtitle = "Lag", ...) }

\arguments{ \item{X}{ }\item{main}{ }\item{xtitle}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
