\name{plot.sme}
\alias{plot.sme}

\title{Plotting function for Sample Mean Excess class}

\description{}

\usage{plot.sme(X, main = attr(X, "desc"), xtitle = get.col.names(attr(X, "data")), ...) }

\arguments{ \item{X}{ }\item{main}{ }\item{xtitle}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
