\name{plotret}
\alias{plot.ret}
\title{Plot Returns}
\description{Plot function for Returns}
\usage{
\method{plot}{ret}(x, style = c("line", "bar"), xlabels = rownames(x), theme.params =
getCurrentTheme(), ...) }
\arguments{ \item{x}{ x}\item{style}{ style}\item{xlabels}{ xlabels}\item{theme.params}{
theme.params}\item{...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
