\name{plot.roi}
\alias{plot.roi}
\title{Plot Return on Investment objects}
\description{Plotting function for Return on Investment objects}
\usage{
\method{plot}{roi}(x, main = "Historical Return on Investment", xtitle = "Lag", ...) 
}
\arguments{ \item{x}{ x}\item{main}{ main}\item{xtitle}{ xtitle}\item{...}{ Further arguments to or
from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
