\name{ppo}
\alias{ppo}

\title{Percentage Price oscillator}

\description{Compute Percentage Price oscillator (Technical Analysis)}

\usage{ppo(X, fast.lag = 10, slow.lag = 30, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{fast.lag}{ }\item{slow.lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
