\name{pptfopt}
\alias{print.PtfOpt}
\title{Print Optimum Portfolio}
\description{Print function for Optimum Ptf}
\usage{
\method{print}{PtfOpt}(x, ...) 
}

\arguments{ 
\item{x}{ OBJECT of class "PrfOpt".}
\item{...}{ Further arguments to or from other methods}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
