\name{prbsar}
\alias{prbsar}

\title{Parabolic Stop and Reverse (PSAR)}

\description{Compute Parabolic Stop and Reverse (PSAR) (Technical Analysis)}

\usage{prbsar(Close, High, Low, accel = c(0.02, 0.2), plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{accel}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
