\name{Pred.Garch}
\alias{Pred.Garch}

\title{Predict Garch model}

\description{Predict Garch model}

\usage{Pred.Garch(X, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
