\name{Pred.VecAr}
\alias{Pred.VecAr}

\title{VAR predictions}

\description{Predict VAR model}

\usage{Pred.VecAr(X, steps = 5, CI = 0.95, viewby = c("vars", "step")) }

\arguments{ \item{X}{ }\item{steps}{ }\item{CI}{ }\item{viewby}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
