\name{pred_error}
\alias{pred_error}

\title{Prediction error}

\description{Calculate Prediction error for model evaluation }

\usage{pred_error(actual, predicted, pc = FALSE) }

\arguments{ \item{actual}{ }\item{predicted}{ }\item{pc}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
