\name{predict.mreg}
\alias{predict.mreg}

\title{Predict method for Multiple regressions}

\description{Predict function for class 'mreg'}

\usage{predict.mreg(mod, ...) }

\arguments{ \item{mod}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
