\name{predict.reg}
\alias{predict.reg}

\title{Predict method for regression}

\description{Predict function for class 'reg'}

\usage{predict.reg(mod, newdata = NULL, ci = 0.95, mode = c("response", ""),
plot = FALSE, shaded = FALSE, xlabels = NULL, main = "Linear Model
Prediction", col = getThemeAttr("col", exact = TRUE)[c(1, 2, 2)],
shade.stripes = 1, shade.col = getThemeAttr("col", exact = TRUE)[2],
shade.density = 40, shade.angle = 30, legend = NULL, ...) }

\arguments{ \item{mod}{ }\item{newdata}{ }\item{ci}{ }\item{mode}{ }\item{plot}{ }\item{shaded}{ }\item{xlabels}{ }\item{main}{ }\item{col}{ }\item{shade.stripes}{ }\item{shade.col}{ }\item{shade.density}{ }\item{shade.angle}{ }\item{legend}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
