\name{predvear}

\alias{predict.VecAr}

\title{VAR predictions}

\description{Predict VAR model}

\usage{
\method{predict}{VecAr}(object, steps = 5, CI = 0.95, viewby = c("vars", "step"), ...) 
}

\arguments{ \item{object}{ OBJECT of class "VecAr".}\item{steps}{ steps}\item{CI}{ CI}\item{viewby}{
viewby}
\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}



