\name{print.Factorise}

\alias{print.Factorise}

\title{Print "factorise"}

\description{Print method for "factorise"}

\usage{print.Factorise(x, ...) }

\arguments{\item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
