\name{print.fft}
\alias{print.fft}

\title{Print FFT}

\description{Print function for 'FFT'}

\usage{print.fft(Xf) }

\arguments{ \item{Xf}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
