\name{print.fs}
\alias{print.fs}

\title{Printing function for Financial Series class}

\description{}

\usage{print.fs(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
