\name{print.Garch}
\alias{print.Garch}

\title{Print Garch}

\description{Print function for Garch model}

\usage{print.Garch(x, digits = 5) }

\arguments{ \item{x}{ }\item{digits}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
