\name{print.GrangCas}
\alias{print.GrangCas}

\title{Print Granger test}

\description{Print function for Granger test}

\usage{print.GrangCas(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
