\name{print.mreg}
\alias{print.mreg}

\title{Print function for mreg}

\description{Print function for class 'mreg'}

\usage{print.mreg(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
