\name{print.PtfOpt}
\alias{print.PtfOpt}

\title{Print Optimum Portfolio}

\description{Print function for Optimum Ptf}

\usage{print.PtfOpt(x) }

\arguments{ \item{x}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
