\name{print.reg}
\alias{print.reg}

\title{Print function for reg}

\description{Print function for class 'reg'}

\usage{print.reg(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
