\name{print.sme}
\alias{print.sme}

\title{Printing function for Sample Mean Excess class}

\description{}

\usage{print.sme(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
