\name{print.VecAr}
\alias{print.VecAr}

\title{Print VAR}

\description{Print function for VAR}

\usage{print.VecAr(x, ...) }

\arguments{ \item{x}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
