\name{printfft}

\alias{print.FFT}

\title{Print FFT results}

\description{Print method for class 'FFT'}

\usage{
\method{print}{FFT}(x, ...)
}

\arguments{
\item{x}{ Instance of class 'FFT'}
\item{...}{ Further arguments to and from other methods}
}

\value{Void}

\author{RAdamant Development Team \email{team@r-adamant.org}}
