\name{printfs}

\alias{print.fs}

\title{Print fs data}

\description{Print method for Financial Series (fs) object.}

\usage{
\method{print}{fs}(x, ...)
}

\arguments{
\item{x}{ Instance of class 'fs'}
\item{...}{Not Used. For compatibility with the generics print function.}
}

\value{Void}

\author{RAdamant Development Team \email{team@r-adamant.org}}
