\name{printvar}

\alias{print.VaR}

\title{Print VaR results}

\description{Print method for class 'VaR'}

\usage{
\method{print}{VaR}(x, ...)
}

\arguments{
\item{x}{ Instance of class 'VaR'}
\item{...}{ Further arguments to and from other methods}
}

\value{Void}

\author{RAdamant Development Team \email{team@r-adamant.org}}


\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
