\name{pro}
\alias{pro}
\title{Price oscillator}
\description{Compute Price oscillator (Technical Analysis)}
\usage{pro(Close, fast.lag = 5, slow.lag = 10, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{fast.lag}{ fast.lag}\item{slow.lag}{
slow.lag}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}

\details{The Price Oscillator Percent shows the percentage difference between two moving averages. A
buy signal is generate when the Price Oscillator Percent rises above zero, and a sell signal when
the it falls below zero.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
