\name{ProbHit}
\alias{ProbHit}

\title{Probability of first hit (Brownian motion)}

\description{Calcualte probability to Hit a barrier}

\usage{ProbHit(B = 0, S0 = 0, mi, sigma) }

\arguments{ \item{B}{ }\item{S0}{ }\item{mi}{ }\item{sigma}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
