\name{psme}
\alias{print.sme}
\title{Print Sample Mean Excess class}
\description{Printing function for Sample Mean Excess class}
\usage{
\method{print}{sme}(x, ...) 
}
\arguments{ \item{x}{ OBJECT of class "sme".}\item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
