\name{PtfBeta}
\alias{PtfBeta}
\title{Portfolio Beta}
\description{Get portfolio Beta}
\usage{PtfBeta(beta, w = NULL, glob = TRUE) }
\arguments{\item{beta}{ beta}\item{w}{ w}\item{glob}{ glob}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
