\name{ptfopt}
\alias{PtfOpt}
\title{Mean-Variance optimum portfolio}
\description{Calculate mean-variance efficient portfolio}
\usage{PtfOpt(ret = NULL, ptf = NULL, mi = NULL, SIGMA = NULL, volatility = TRUE, ...) }

\arguments{\item{ret}{ ret}\item{ptf}{ ptf}\item{mi}{ mi}\item{SIGMA}{ SIGMA}\item{volatility}{
volatility}\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
