\name{PtfRet}
\alias{PtfRet}

\title{Portfolio returns}

\description{Calculate portfolio returns}

\usage{PtfRet(PTF, w = NULL, glob = TRUE, calc.ret = FALSE, ...) }

\arguments{ \item{PTF}{ }\item{w}{ }\item{glob}{ }\item{calc.ret}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
