\name{PtfUtility}
\alias{PtfUtility}

\title{Portfolio Utility}

\description{Calculate utility and plot for efficient portfolio}

\usage{PtfUtility(PTF = NULL, W, R = NULL, SIGMA = NULL, af = 3, plot = TRUE, plot.mv = FALSE, plot.mu = FALSE, ...) }

\arguments{ \item{PTF}{ }\item{W}{ }\item{R}{ }\item{SIGMA}{ }\item{af}{ }\item{plot}{ }\item{plot.mv}{ }\item{plot.mu}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
