\name{PtfVar}
\alias{PtfVar}
\title{Portfolio variance}
\description{Calculate portfolio variance}

\usage{PtfVar(PTF, w = NULL, glob = TRUE, 
vol = FALSE, calc.ret = FALSE, ...) }

\arguments{\item{PTF}{ PTF}\item{w}{ w}\item{glob}{ glob}\item{vol}{ vol}\item{calc.ret}{
calc.ret}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
