\name{print.VecAr}

\alias{print.VecAr}

\title{Print VAR}

\description{Print method for VAR}

\usage{
\method{print}{VecAr}(x, ...) 
}

\arguments{ \item{x}{ OBJECT of class "VecAr".}\item{...}{ Further arguments to or from other
methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
