\name{pvt}
\alias{pvt}
\title{Price Volume trend indicator}
\description{Compute Price Volume trend indicator (Technical Analysis)}
\usage{pvt(Close, Volume, lag = 5, plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{Volume}{Vector. Asset traded
Volume.}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}



\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


