\name{qgev}
\alias{qgev}

\title{Generalised Extreme Value (GEV) - Quantile function}

\description{}

\usage{qgev(P, mu = 0, xi = 0.1, sigma = 1) }

\arguments{ \item{P}{ }\item{mu}{ }\item{xi}{ }\item{sigma}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
