\name{qgpd}
\alias{qgpd}

\title{Generalised Pareto Distribution (GPD) - Quantile function}

\description{}

\usage{qgpd(P, xi = 0.1, sigma = 1, trsh = 0) }

\arguments{ \item{P}{ }\item{xi}{ }\item{sigma}{ }\item{trsh}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
