\name{RAdamant}
\alias{RAdamant}
\alias{RAdamant}
\docType{package}
\title{
Financial Technical Analysis and Risk Management
}
\description{
R-Adamant is a collection of functions and algorithms for processing of Financial Time Series, Risk Management and Econometrics.}

\details{

\tabular{ll}{
Package: \tab RAdamant \cr
Type: \tab Package \cr
Version: \tab 0.7.1 \cr
Date: \tab 2011-06-25 \cr
License: \tab GPL>=2 \cr
LazyLoad: \tab yes \cr
}

}

\author{
RAdamant Development Team

Maintainer: RAdamant Development Team <team@r-adamant.org>
}

\references{
Farewell friend, I was a thousand times more evil than thou...
}

\keyword{package}

\seealso{}

\examples{}
