\name{radpkg}
\alias{RAdamant}
\alias{RAdamant}
\docType{package}
\title{
Financial Technical Analysis and Risk Management
}
\description{
R-Adamant is a collection of functions and algorithms for processing of Financial Time Series, Risk
Management and Econometrics.}
\details{
\tabular{ll}{
Package: \tab RAdamant \cr
Type: \tab Package \cr
Version: \tab 0.8.2 \cr
Date: \tab 2011-07-17 \cr
License: \tab GPL>=2 \cr
LazyLoad: \tab yes \cr
}
}
\author{
RAdamant Development Team
Maintainer: RAdamant Development Team <team@r-adamant.org>
}

\keyword{package}
