\name{recycle}
\alias{recycle}

\title{Recycle function for time series}

\description{Recycle an input sequence X to get a new sequence of the specified length V}

\usage{recycle(X, V = length(X)) }

\arguments{ \item{X}{ }\item{V}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
