\name{RelVol}
\alias{RelVol}

\title{Relative Volatility oscillator}

\description{Compute Relative Volatility oscillator (Technical Analysis)}

\usage{RelVol(Close, sdlag = 9, lag = 5) }

\arguments{ \item{Close}{ }\item{sdlag}{ }\item{lag}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
