\name{rgev}
\alias{rgev}
\title{Generalised Extreme Value (GEV) - Random Numbers Generator}
\description{Generalised Extreme Value (GEV) - Random Numbers Generator}
\usage{rgev(N, mu = 0, xi = 0.1, sigma = 1) }
\arguments{
\item{N}{ N}
\item{mu}{ mu}
\item{xi}{ xi}
\item{sigma}{ sigma}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


