\name{rgpd}
\alias{rgpd}

\title{Generalised Pareto Distribution (GPD) - Random Numbers Generator}

\description{}

\usage{rgpd(n, xi = 0.1, sigma = 1, trsh = 0) }

\arguments{ \item{n}{ }\item{xi}{ }\item{sigma}{ }\item{trsh}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
