\name{roc}
\alias{roc}
\title{Rate of Change index}
\description{Compute Rate of Change index (Technical Analysis)}
\usage{roc(X, lag = 5, pc = TRUE, plot = TRUE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{pc}{ pc}\item{plot}{
LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}

\details{The Rate of Change function measures rate of change relative to previous periods. The
function is used to determine how rapidly the data is changing. The factor is usually 100, and is
used merely to make the numbers easier to interpret or graph. The function can be used to measure
the Rate of Change of any data series, such as price or another indicator. When used with the price,
it is referred to as the Price Rate Of Change, or PROC.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
