\name{ROCplot}

\alias{ROCplot}
\alias{ROCplot.scorecard}

\title{ROC analysis}

\description{Plot ROC curve}

\usage{
ROCplot(x, ...) \cr
ROCplot.scorecard(sc) 
}

\arguments{
\item{sc}{ }
\item{x}{ }
\item{...}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
