\name{root.search.interval}
\alias{root.search.interval}

\title{Compute a proper search interval for uniroot function}

\description{}

\usage{root.search.interval(from, func = NULL, type = c("left", "both", "right"), max.iter = 500, show.warnings = FALSE, debug = FALSE, ...) }

\arguments{ \item{from}{ }\item{func}{ }\item{type}{ }\item{max.iter}{ }\item{show.warnings}{ }\item{debug}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
