\name{RowMax}
\alias{rowMax}
\alias{rowMin}

\title{Maximum / Minimum by row}

\description{
\code{rowMax}: Compute parallel max across the rows of X \cr
\code{rowMin}: Compute parallel min across the rows of X 
}

\usage{
rowMax(X) \cr
rowMin(X)
}

\arguments{ 
\item{X}{ Input matrix/sequence}
}

\details{}

\value{A matrix NROW(X) by one, where each row is the max / min of the rows of X).}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
