\name{rschint}
\alias{root.search.interval}
\title{Interval for uniroot function}
\description{Compute a proper search interval for uniroot function}

\usage{root.search.interval(from, func = NULL, 
type = c("left", "both", "right"), max.iter = 500,
show.warnings = FALSE, debug = FALSE, ...) }

\arguments{
\item{from}{ from}
\item{func}{ func}
\item{type}{ type}
\item{max.iter}{
max.iter}
\item{show.warnings}{ show.warnings}
\item{debug}{ debug}
\item{...}{ Further arguments to or
from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


