\name{rsi}
\alias{rsi}

\title{Relative strength indicator}

\description{Compute Relative strength indicator (Technical Analysis)}

\usage{rsi(X, lag, plot = FALSE, ...) }

\arguments{
\item{X}{ X}
\item{lag}{ INTEGER. Number of lag periods.}
\item{plot}{ LOGICAL. If TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods.}
}

\details{The Relative Strength Index (RSI) is an oscillator that measures current price strength in
relation to previous prices. \cr The RSI is a versatile tool, it can be used to generate buy and
sell signals, show overbought and oversold conditions, confirm price movement and warn of potential
price reversals through divergences.}


\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
