\name{rvi}
\alias{rvi}
\title{Relative Vigor indicator}
\description{Compute Relative Vigor indicator (Technical Analysis)}
\usage{
rvi(Close
   , High = NULL
   , Low = NULL
   , Open = NULL
   , plot = TRUE
   , ...
   )
}
\arguments{
\item{ Close}{Vector. Close price.}
\item{ High}{Vector. High price.}
\item{ Low}{Vector. Low price.}
\item{ Open}{Vector. Open price.}
\item{ plot}{Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


