\name{SampMom}
\alias{SampMom}

\title{Sample moments (Brownian motion)}

\description{Calculate sample moments of a Brownian motion}

\usage{SampMom(P, X, moms = 1:2) }

\arguments{ \item{P}{ }\item{X}{ }\item{moms}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
