\name{Score.card}

\alias{Score.card}

\title{Score Card}

\description{Create Credit Score Card based on Logistic Regression }

\usage{Score.card(Y, X, woe) }

\arguments{\item{Y}{ }\item{X}{ }\item{woe}{ }}

\details{For the moment only numerical variables! Work in progress...}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
