\name{Score.card}
\alias{Score.card}
\title{Score Card}
\description{Create Credit Score Card based on Logistic Regression }
\usage{Score.card(Y, X, nseg = 2) }


\arguments{
\item{Y}{ VECTOR. Target variable in 0-1 format.}
\item{X}{ DATA.FRAME / MATRIS of regressors.} 
\item{nseg}{ INTEGER / VECTOR. Number of segments to factorise numerical variables.}}

\details{For the moment only numerical variables! Work in progress...}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
