\name{sde}
\alias{sde}
\title{Error standard deviation}
\description{Calculate Error standard deviation for model evaluation }
\usage{sde(actual, predicted) }
\arguments{\item{actual}{ actual}\item{predicted}{ predicted}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
