\name{sensAnalysis.default}
\alias{sensAnalysis.default}

\title{Sensitivity analysis – default method}

\description{}

\usage{sensAnalysis.default(X, win.size = length(coef(X)), plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{win.size}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
