\name{sensAnalysis.lm}
\alias{sensAnalysis.lm}

\title{Sensitivity analysis – method for lm}

\description{}

\usage{sensAnalysis.lm(X, ...) }

\arguments{ \item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
