\name{sensAnalysis}
\alias{sensAnalysis}

\title{Sensitivity analysis – method}

\description{}

\usage{sensAnalysis(X, ...) }

\arguments{ \item{X}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
