\name{sensanlm}
\alias{sensAnalysis.lm}

\title{Sensitivity analysis method for lm}

\description{Sensitivity analysis method for lm}

\usage{
\method{sensAnalysis}{lm}(X, ...) 
}

\arguments{ \item{X}{ OBJECT of class "lm".}
\item{...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


