\name{setProjectionMatrix}
\alias{setProjectionMatrix}

\title{Save 3D Projection matrix}

\description{Plotting tools}

\usage{setProjectionMatrix(pmat = NULL, env = getOption("RAdamant")) }

\arguments{ \item{pmat}{ }\item{env}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
