\name{shade.plot}
\alias{shade.plot}

\title{Graphics parameters function}

\description{Plotting tools}

\usage{shade.plot(to = NULL, from = NULL, base = NULL, theme.params = getCurrentTheme(), overrides = list(...), ...) }

\arguments{ \item{to}{ }\item{from}{ }\item{base}{ }\item{theme.params}{ }\item{overrides}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
