\name{Sharpe}
\alias{Sharpe}
\alias{Sharpe.default}
\alias{Sharpe.Capm}

\title{Sharpe index}

\description{
\code{Sharpe}: Calculate Sharpe index for a portfolio. \cr
\code{Sharpe.Capm}: Get Sharpe index from an object of class. "Capm"
}

\usage{
Sharpe(x, ...)

\method{Sharpe}{default}(PTF, rfr = 0, ...) 

\method{Sharpe}{Capm}(X, rfr = 0, ...) 

}

\arguments{ 
\item{PTF}{ Input portfolio}
\item{rfr}{ risk free rate}
\item{X}{An object of class "Capm"}
\item{...}{ }}

\details{}

\value{}

\references{}

\author{ RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{\code{\link{Treynor}}, \code{\link{Jensen}}, \code{\link{Appraisal}}}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
