\name{SI.format}
\alias{SI.format}

\title{Number conversion}

\description{Plotting tools}

\usage{SI.format(x) }

\arguments{ \item{x}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
