\name{sme}
\alias{sme}

\title{Sample Mean Excess function}

\description{}

\usage{sme(X, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
