\name{SomerD}

\alias{SomerD}

\title{Somers D}

\description{Calculate Somers d statistic}

\usage{SomerD(target, pred, ...) }

\arguments{\item{target}{ }\item{pred}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
