\name{sort}

\alias{SORT}

\title{Sort matrix}

\description{Sort each column of the input matrix X independently}

\usage{SORT(x, decreasing = FALSE, ...)}

\arguments{
\item{x}{ Input matrix.}
\item{decreasing}{ LOGICAL. Decreasing order.}
\item{...}{ Further arguments to or from other methods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
