\name{specgram}
\alias{specgram}
\title{Spectrogram}
\description{Create spectrogram for FFT}
\usage{specgram(X, win.size = max(1, NROW(X)/20), plot = TRUE, ...) }
\arguments{\item{X}{ X}\item{win.size}{ win.size}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
