\name{splitWindow}
\alias{splitWindow}
\title{Sliding windows}
\description{Sliding windows}
\usage{splitWindow(N, direction = c("forward", "backward"), 
mode = c("EW", "SW"), from = NULL,
win.size = 1, by = 1, labels = 1:N, ...) }
\arguments{\item{N}{ N}\item{direction}{ direction}\item{mode}{ mode}\item{from}{
from}\item{win.size}{ win.size}\item{by}{ by}\item{labels}{ labels}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
