\name{starc}
\alias{starc}
\title{Stoller Starc bands}
\description{Compute Stoller Starc bands (Technical Analysis)}
\usage{starc(Close, High = NULL, Low = NULL, atr.mult = 2, lag = 5, atr.lag =
14, mov = c("sma", "ema", "wma"), plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{atr.mult}{ atr.mult}
\item{lag}{Integer. Number of lag periods.}
\item{atr.lag}{
atr.lag}
\item{mov}{ mov}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


