\name{statusbar}
\alias{statusbar}

\title{Interactive status bar for console logging}

\description{}

\usage{statusbar(message = "Computing..", status = 0, n = 1, N = 1, step = 0.01) }

\arguments{ \item{message}{ }\item{status}{ }\item{n}{ }\item{N}{ }\item{step}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
