\name{StepMat}
\alias{StepMat}

\title{Step matrix for binomial tree}

\description{Simulate binomial path of a binomial tree}

\usage{StepMat(init, n_step, up, down) }

\arguments{ \item{init}{ }\item{n_step}{ }\item{up}{ }\item{down}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
