\name{Strvar}
\alias{Strvar}

\title{Structural Vector Autoregressive model}

\description{Estimate Structural Vector Autoregressive model}

\usage{Strvar(X, A = NULL, B = NULL, inter = FALSE) }

\arguments{ \item{X}{ }\item{A}{ }\item{B}{ }\item{inter}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
