\name{Styles}
\alias{Styles}
\title{Styles analysis (portfolio)}
\description{Perform Style analysis  }
\usage{Styles(FUND, IND, W, lower = NULL, upper = NULL, ...) }
\arguments{\item{FUND}{ FUND}\item{IND}{ IND}\item{W}{ W}\item{lower}{ lower}\item{upper}{
upper}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
