\name{Sum.dens}
\alias{Sum.dens}

\title{Plot summary information}

\description{Plot summary information of a vector with its density}

\usage{Sum.dens(x) }

\arguments{ 
\item{x}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
