\name{SumClust}
\alias{SumClust}

\title{Summary clusters}

\description{Summary function for time series clusters}

\usage{SumClust(cluster, funs = summary, ...) }

\arguments{ \item{cluster}{ }\item{funs}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
