\name{sumdens}
\alias{Sum.dens}
\title{Plot summary information}
\description{Plot summary information of a vector with its density}
\usage{Sum.dens(x, ...) }
\arguments{
\item{x}{ VECTOR. Input series.}
\item{...}{ further arguments for "plot" function}
}
\author{RAdamant Development Team \email{team@r-adamant.org} }
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
