\name{summary.mreg}
\alias{summary.mreg}

\title{Summary method for mreg}

\description{}

\usage{summary.mreg(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
