\name{summary.reg}
\alias{summary.reg}

\title{Summary method for reg}

\description{}

\usage{summary.reg(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
