\name{summary.scorecard}

\alias{summary.scorecard}

\title{Summary Scorecard}

\description{Summary method for "Scorecard" object}

\usage{summary.scorecard(sc, plot = FALSE, ...) }

\arguments{\item{sc}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
