\name{summary.univar}
\alias{summary.univar}

\title{Summary method for univariate analysis}

\description{}

\usage{summary.univar(X) }

\arguments{ \item{X}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
