\name{SummaryDD}
\alias{SummaryDD}

\title{Summary function for drawdown}

\description{}

\usage{SummaryDD(DD) }

\arguments{ \item{DD}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
