\name{sumreg}
\alias{summary.mreg}
\title{Summary method for mreg}
\description{Summary method for mreg}

\usage{
\method{summary}{mreg}(object, ...) }

\arguments{ 
\item{object}{ OBJECT of class "mreg"}
\item{...}{ Further arguments to or from other methods}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
