\name{sumscrd}

\alias{summary.scorecard}

\title{Summary Scorecard}

\description{Summary method for "Scorecard" object}

\usage{
\method{summary}{scorecard}(object, plot = FALSE, ...) 
}

\arguments{\item{object}{ OBJECT of class "scorecard".}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
