\name{swing}
\alias{Swing}
\title{Swing Index}
\description{Calculate Swing index (Technical Analysis)}
\usage{Swing(Close, High, Low, Open, ret_cum = FALSE, plot = FALSE, ...)}
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{Open}{Vector. Open price.}
\item{ret_cum}{ ret_cum}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


