\name{symbol.lookup}
\alias{symbol.lookup}

\title{Lookup Stock Symbol from Yahoo!}

\description{}

\usage{symbol.lookup(what = \) }

\arguments{ \item{what}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
