\name{symlkup}
\alias{symbol.lookup}
\title{Lookup Stock Symbol from Yahoo!}
\description{Lookup Stock Symbol from Yahoo!}
\usage{symbol.lookup(what = "") }
\arguments{ \item{what}{ what}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
