\name{thigh}
\alias{thigh}

\title{True High oscillator}

\description{Compute True High oscillator (Technical Analysis)}

\usage{thigh(Close, High = NULL, lag = 5, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
