\name{tirLev}
\alias{tirLev}

\title{Trione levels}

\description{Compute Trione levels (Technical Analysis)}

\usage{tirLev(High, Low, lag = 5, plot = FALSE, ...) }

\arguments{ \item{High}{ }\item{Low}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
