\name{tlow}
\alias{tlow}

\title{True Low oscillator}

\description{Compute True Low oscillator (Technical Analysis)}

\usage{tlow(Close, Low = NULL, lag = 5, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{Low}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
