\name{trf}
\alias{trf}
\title{(Average) True range}
\description{Compute (Average) True range (Technical Analysis)}
\usage{trf(Close, High = NULL, Low = NULL, lag = 1, 
average = TRUE, avg.lag = 14, plot = FALSE, ...)
}
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{lag}{Integer. Number of lag periods.}
\item{average}{ average}
\item{avg.lag}{
avg.lag}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


