\name{track_sign}
\alias{track_sign}
\title{Track signal}
\description{Calculate Track signal for model evaluation }
\usage{track_sign(actual, predicted) }
\arguments{\item{actual}{ actual}\item{predicted}{ predicted}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
