\name{typ}
\alias{tyP}
\title{Typical price}
\description{Compute Typical price (Technical Analysis)}
\usage{tyP(Close, High, Low, plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


