\name{ulcer}
\alias{ulcer}

\title{Ulcer index}

\description{Compute Ulcer index (Technical Analysis)}

\usage{ulcer(X, lag, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
