\name{ultima}
\alias{ultima}
\title{Ultima oscillator}
\description{Compute Ultima oscillator (Technical Analysis)}

\usage{ultima(Close
	, High = NULL
	, Low = NULL
	, lag = 1
	, win1 = 7
	, win2 = 14
	, win3 = 28
	, plot = TRUE
	, ...) 
}

\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{lag}{Integer. Number of lag periods.}
\item{win1}{ win1}
\item{win2}{
win2}
\item{win3}{ win3}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments
to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


