\name{vcmof}
\alias{vcmof}
\title{Variable Chande Momentum Oscillator}
\description{Compute Variable Chande Momentum Oscillator (Technical Analysis)}
\usage{vcmof(X, lag = 5, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


