\name{vhff}
\alias{vhff}
\title{Vertical Horizontal Filter}
\description{Compute Vertical Horizontal Filter (Technical Analysis)}
\usage{vhff(X, lag = 9, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
