\name{wad}
\alias{wad}

\title{Williams Advance Decline}

\description{Compute Williams Advance Decline (Technical Analysis)}

\usage{wad(Close, High = NULL, Low = NULL, lag = 5, na.rm = FALSE, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{lag}{ }\item{na.rm}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
