\name{weigevid}
\alias{WeightEvid}
\title{Weight of Evidence}
\description{Calculate weight of evidence for a matrix with target variable}
\usage{WeightEvid(data, target, nseg, missing = FALSE, na.replace = NULL) }
\arguments{\item{data}{ data}\item{target}{ target}\item{nseg}{ nseg}\item{missing}{
missing}\item{na.replace}{ na.replace}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
