\name{WeightEvid}

\alias{WeightEvid}

\title{Weight of Evidence}

\description{Calculate weight of evidence for a matrix with target variable}

\usage{WeightEvid(data, target, nseg, missing = FALSE, na.replace = NULL) }

\arguments{\item{data}{ }\item{target}{ }\item{nseg}{ }\item{missing}{ }\item{na.replace}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
