\name{wildavg}
\alias{wildAvg}
\title{Wilder Moving Average}
\description{Compute Wilder Moving Average (Technical Analysis)}
\usage{wildAvg(X, lag = 5, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ lag}\item{plot}{ plot}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
