\name{wildSum}
\alias{wildSum}

\title{Wilder Summation}

\description{Compute Wilder Summation (Technical Analysis)}

\usage{wildSum(x, lag = 5) }

\arguments{ \item{x}{ }\item{lag}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
