\name{wro}
\alias{wro}
\title{Williams R}
\description{Compute Williams R (Technical Analysis)}
\usage{wro(Close, High = NULL, Low = NULL, lag = 5, plot = TRUE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


