\name{Zind}
\alias{Zind}

\title{Z index}

\description{Compute the Z-score of X (Standardize each column of X)}

\usage{Zind(x, sigma = 1, mi = 2) }

\arguments{ \item{x}{ }\item{sigma}{ }\item{mi}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
