\name{abi}

\alias{Abi}

\title{Absolute Breath Index - ABI}

\description{Compute Absolute Breath Index (Technical Analysis)}

\usage{Abi(X, lag = 5, plot=FALSE, ...)}

\arguments{ 
\item{X}{ Input numerical series}
\item{lag}{ Number of lags}
\item{plot}{ Logical. Return plot.}
\item{...}{ Further arguments to or from other methods}
}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


