\name{absrs}
\alias{absrs}

\title{Absolute Relative Strenght}

\description{Compute Absolute Relative Strenght (Technical Analysis)}

\usage{absrs(X, lag = 14, na.rm = FALSE, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{na.rm}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
