\name{absrs}
\alias{absrs}
\title{Absolute Relative Strenght}
\description{Compute Absolute Relative Strenght (Technical Analysis)}
\usage{absrs(X, lag = 14, na.rm = FALSE, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{na.rm}{ na.rm}\item{plot}{
LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


