\name{Accuracy}
\alias{accuracy}
\alias{accuracy.scorecard}
\title{Accuracy}
\description{Get accuracy measure from the results of a classification model.}

\usage{
accuracy(x, ...) 
\method{accuracy}{scorecard}(x, th, ...) 
}

\arguments{
\item{x}{ x}
\item{th}{ th}
\item{...}{ Further arguments to or from other methods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
