\name{acdi}
\alias{acdi}
\title{Acceleration Deceleration}
\description{Acceleration Deceleration Technical Indicator}

\usage{acdi(Close, High = NULL, Low = NULL, Vol = NULL, plot = TRUE, ...)}

\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Vol}{ VECTOR. Asset traded Volume.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}



\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


