\name{acdi}
\alias{acdi}
\title{Acceleration Deceleration}
\description{Acceleration Deceleration Technical Indicator}

\usage{
acdi(Close
    , High = NULL
    , Low = NULL
    , Vol = NULL
    , plot = TRUE
    , ...
    )
}

\arguments{
\item{Close}{ Vector. Close price.}
\item{ High}{ Vector. High price.}
\item{ Low}{ Vector. Low price.}
\item{ Vol}{ Vector. Asset traded Volume.}
\item{ plot}{ Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


