\name{adi}
\alias{ADind}
\title{Advance-Decline Indicator}
\description{Advance-Decline Indicator (Technical Analysis)}
\usage{ADind(close, high, low, lag = 5)}
\arguments{
\item{close}{ VECTOR. Close price.}
\item{high}{ VECTOR. high price.}
\item{low}{ VECTOR. Low price.}
\item{lag}{ INTEGER. Number of lag periods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


