\name{adi}
\alias{ADind}
\title{Advance-Decline Indicator}
\description{Advance-Decline Indicator (Technical Analysis)}
\usage{ADind(close, high, low, lag = 5)}
\arguments{
\item{close}{ Vector. Close price.}
\item{high}{ Vector. high price.}
\item{low}{ Vector. Low price.}
\item{lag}{ Integer. Number of lag periods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


