\name{adrating}

\alias{ADrating}

\title{Average Directional Rating }

\description{Compute Average Directional Rating index (Technical Analysis)}

\usage{ADrating(close, high, low, lag)}

\arguments{
\item{close}{ VECTOR. Close price.}
\item{high}{ VECTOR. high price.}
\item{low}{ VECTOR. Low price.}
\item{lag}{ INTEGER. Number of lag periods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


