\name{adratio}
\alias{ADratio}

\title{Advance Decline ratio}

\description{Compute Advance Decline ratio (Technical Analysis)}

\usage{ADratio(X, lag, plot, ...) }

\arguments{ 
\item{X}{ X}
\item{ lag}{ Integer. Number of lag periods.}
\item{ plot}{ Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}}


\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


