\name{ADratio}
\alias{ADratio}

\title{Advance Decline ratio}

\description{Compute Advance Decline ratio (Technical Analysis)}

\usage{ADratio(X, lag, plot, ...) }

\arguments{ \item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
