\name{AdvDec}
\alias{AdvDec}

\title{Advance Decline issues}

\description{Compute Advance Decline issues (Technical Analysis)}

\usage{AdvDec(X, lag = 5, ret.idx = TRUE, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{ret.idx}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
