\name{advdec}
\alias{AdvDec}
\title{Advance Decline issues}
\description{Compute Advance Decline issues (Technical Analysis)}
\usage{AdvDec(X, lag = 5, ret.idx = TRUE, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{ret.idx}{
ret.idx}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


