\name{empty}
\alias{empty}

\title{empty}

\description{empty}

\usage{empty}

\arguments{ empty}

\details{An indicator based on the difference between two exponential moving averages, expressed in
absolute terms.  \cr Also known as the MACD indicator, the APO is calculated by subtracting the
longer exponential moving average from the shorter exponential moving average.}

\value{empty}

\references{empty}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{empty}

\examples{empty}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
