\name{apo}
\alias{apo}
\title{Apo - Absolute price indicator}
\description{Apo - Absolute price indicator}
\usage{apo(X, fast.lag = 10, slow.lag = 30, plot = FALSE, ...)}
\arguments{\item{X}{ X}\item{fast.lag}{ fast.lag}\item{slow.lag}{ slow.lag}\item{plot}{ LOGICAL. If
TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


