\name{arms}
\alias{Arms}
\title{Arms index}
\description{Compute Arms index (Technical Analysis)}
\usage{Arms(X, Volume, lag, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{ Volume}{ Vector. Asset traded Volume.}
\item{ lag}{ Integer. Number of lag periods.}
\item{ plot}{ Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


