\name{aroon}
\alias{aroon}
\title{Aroon oscillator}
\description{Compute Aroon oscillator (Technical Analysis)}
\usage{aroon(X, lag = 5, plot = TRUE, ...) }
\arguments{
\item{X}{ X}
\item{ lag}{ Integer. Number of lag periods.}
\item{ plot}{ Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


