\name{aroup}
\alias{aroup}
\title{Aroon Up oscillator}
\description{Compute Aroon Up oscillator (Technical Analysis)}
\usage{aroup(X, lag = 5, plot = TRUE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


