\name{asfs}
\alias{as.fs}
\title{Convert Yahoo! Data into Financial Series object}
\description{Convert Yahoo! Data into Financial Series object}
\usage{as.fs(X, SName = "", Symbol = "") }
\arguments{\item{X}{ X}\item{SName}{ SName}\item{Symbol}{ Symbol}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
