\name{bolband}
\alias{BolBand}
\title{Bollinger Bands}
\description{Compute Bollinger Bands (Technical Analysis)}
\usage{
BolBand(Close, High, Low
    , fact = 2
    , win.size = 5
    , plot = FALSE
    , ...
    ) 
}
\arguments{
\item{Close}{ Vector. Close price.}
\item{   High}{ Vector. High price.}
\item{   Low}{ Vector.Low price.}
\item{   fact}{ fact}
\item{   win.size}{ win.size}
\item{   plot}{ Logical. If TRUE plot isreturned.}
\item{   ...}{ Further arguments to or from other methods.}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


