\name{BolBand}
\alias{BolBand}
\title{Bollinger Bands}
\description{Compute Bollinger Bands (Technical Analysis)}
\usage{BolBand(Close, High, Low, fact = 2, win.size = 5, plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{fact}{ fact}\item{win.size}{ win.size}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
