\name{bolbandb}

\alias{BolBandB}

\title{Bollinger Bands Bandwidth}

\description{Compute Bollinger Bands Bandwidth (Technical analysis)}
\usage{
BolBandB(Close, High, Low
    , fact=2
    , win.size=5
    , plot=FALSE
    , ...
    )
}
\arguments{
\item{Close}{ Vector. Close price.}
\item{   High}{ Vector. High price.}
\item{   Low}{ Vector. Low price.}
\item{   fact}{ fact}
\item{   win.size}{ win.size}
\item{   plot}{ Logical. If TRUE plot is returned.}
\item{   ...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


