\name{bolfib}
\alias{Bol.Fib}
\title{Bollinger Bands - Fibonacci ratio}
\description{Compute Bollinger Bands - Fibonacci ratio (Technical Analysis)}
\usage{Bol.Fib(Close, High, Low, win.size = 5, fibo = c(1.618, 2.618, 4.236),
plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{win.size}{ win.size}\item{fibo}{ fibo}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
