\name{boot}
\alias{boot}
\title{General bootstrapping function}
\description{General bootstrapping function}

\usage{
boot(X, nboots = 100, func = NULL, init = NULL, 
message = "Bootstrapping...", ...) 
}

\arguments{\item{X}{ X}\item{nboots}{ nboots}\item{func}{ func}\item{init}{ init}\item{message}{
message}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


