\name{bop}
\alias{Bop}
\title{Balance of Power}
\description{Compute Balance of Power (Technical Analysis)}
\usage{
Bop(Close, Open, High, Low, smoothed = TRUE, ...) 
}
\arguments{
\item{Close}{ Vector. Close price.}
\item{   Open}{ Vector. Open price.}
\item{   High}{ Vector. High price.}
\item{   Low}{ Vector. Low price.}
\item{   smoothed}{ smoothed}
\item{   ...}{ Further arguments to or from other methods.}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


