\name{Bop}
\alias{Bop}
\title{Balance of Power}
\description{Compute Balance of Power (Technical Analysis)}
\usage{Bop(Close, Open, High, Low, smoothed = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Open}{ VECTOR. Open price.}\item{High}{ VECTOR.
High price.}\item{Low}{ VECTOR. Low price.}\item{smoothed}{ smoothed}\item{...}{ Further arguments
to or from other methods.}}
\details{The Balance of Power (BOP) indicator measures the strength of the bulls vs. bears by
assessing the ability of each to push price to an extreme level.\cr The Balance of Power indicator
is calculated by the formula:\cr The resulting Balance of Power value is then typically smoothed by
a moving average.\cr The Balance Of Power indicator can be used to identify the systematic buying
(accumulation) and systematic selling (distribution) of stock shares. \cr BOP does it is a good idea
to conceptualize how much and what type of trading activities take place in the marketplace. More
specifically think of institutional and individual trading activities and the volume of trading that
they do.}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
