\name{box3d}
\alias{box3d}
\title{3D box}
\description{Plotting tools}
\usage{box3d(x, y, z, pmat = getProjectionMatrix(), half = FALSE, ...) }
\arguments{ \item{x}{ X axis}\item{y}{ Y axis}\item{z}{ Z axis}\item{pmat}{ pamt}\item{half}{
half}\item{...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


