\name{bpdlind}
\alias{BPDLind}
\title{BPDL indicator}
\description{Compute BPDL indicator (Technical Analysis)}
\usage{BPDLind(Close, lag = 1, smoothed = TRUE, slag = 5) }
\arguments{
\item{Close}{ Vector. Close price.}
\item{   lag}{ Integer. Number of lag periods.}
\item{   smoothed}{ smoothed}
\item{   slag}{ slag}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


