\name{BPDLind}
\alias{BPDLind}

\title{BPDL indicator}

\description{Compute BPDL indicator (Technical Analysis)}

\usage{BPDLind(Close, lag = 1, smoothed = TRUE, slag = 5) }

\arguments{ \item{Close}{ }\item{lag}{ }\item{smoothed}{ }\item{slag}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
