\name{Breadth}
\alias{Breadth}
\title{Breadth trusth indicator}
\description{Compute Breadth trusth indicator (Technical Analysis)}
\usage{Breadth(X, lag = 5, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}

\details{The Breadth Thrust indicator is a ten period simple moving average of the Advance/Decline
Line Breadth. A
Breadth Thrust formation is defined as when the Breadth Thrust indicator moves from below 40% to
above 61.5 percent
within a ten day period. This indicates that the market is rapidly moving from an oversold condition
to a strong up
trend. A Breadth Thrust formations are rare and are a very good indicator of a bull market.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
