\name{BroMot}
\alias{BroMot}

\title{Browniam motion}

\description{Simulate a standard Brownian motion }

\usage{BroMot(nsim, T, S0 = 0, mi = 0, sigma = 1, geom = TRUE, same.rnd = TRUE, plot = FALSE, ...) }

\arguments{ \item{nsim}{ }\item{T}{ }\item{S0}{ }\item{mi}{ }\item{sigma}{ }\item{geom}{ }\item{same.rnd}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
