\name{bromot}
\alias{BroMot}
\title{Browniam motion}
\description{Simulate a standard Brownian motion }
\usage{
BroMot(nsim, T, S0 = 0, mi = 0, sigma = 1, 
geom = TRUE, same.rnd = TRUE, plot = FALSE, ...) 
}

\arguments{\item{nsim}{ nsim}\item{T}{ T}\item{S0}{ S0}\item{mi}{ mi}\item{sigma}{
sigma}\item{geom}{ geom}\item{same.rnd}{ same.rnd}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
