\name{bromot2d}
\alias{BroMot2D}
\title{2-dimensional Browniam motion}
\description{Simulate a bi-dimensional standard Brownian motion }

\usage{
BroMot2D(nsim, T, S0, mi, sigma, geom = TRUE, 
same.rnd = FALSE, laydisp = NULL, plot = TRUE, ...) 
}
\arguments{\item{nsim}{ nsim}\item{T}{ T}\item{S0}{ S0}\item{mi}{ mi}\item{sigma}{
sigma}\item{geom}{ geom}\item{same.rnd}{ same.rnd}\item{laydisp}{ laydisp}\item{plot}{ LOGICAL. If
TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
