\name{bsmomt}
\alias{BS.moments}
\title{Black & Scholes moments}
\description{Calculate first four moments for Black & Scholes}
\usage{BS.moments(BS = NULL, under, rfr, sigma, yield, maty) }
\arguments{\item{BS}{ BS}\item{under}{ under}\item{rfr}{ rfr}\item{sigma}{ sigma}\item{yield}{
yield}\item{maty}{ maty}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
