\name{buypre}
\alias{buypre}
\title{Buying pressure indicator}
\description{Compute Buying pressure indicator (Technical Analysis)}
\usage{buypre(Close, Low, lag = 5, plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Low}{ VECTOR. Low price.}\item{lag}{ INTEGER.
Number of lag periods.}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


