\name{buypre}
\alias{buypre}

\title{Buying pressure indicator}

\description{Compute Buying pressure indicator (Technical Analysis)}

\usage{buypre(Close, Low, lag = 5, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{Low}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
