\name{capm}

\alias{Capm.default}
\alias{Capm}

\title{Capm - default method}

\description{Default method for CAPM}

\usage{
Capm(PTF, ...)
\method{Capm}{default}(PTF, PTF_M, rf = NULL, rfr = NULL, ...) 
}

\arguments{ 
\item{PTF}{ Prtfolio.}
\item{PTF_M}{ Market portfolio.}
\item{rf}{ Risk free asset.}
\item{rfr}{ Risk free rate.}
\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
