\name{cciv2}
\alias{cci.v2}
\title{Commodity channel index v02}
\description{Compute Commodity channel index v02 (Technical Analysis)}

\usage{cci.v2(High, Low, Close, lag = 5, plot = FALSE, ...) }

\arguments{
\item{High}{ Vector. High price.}
\item{   Low}{ Vector. Low price.}
\item{   Close}{ Vector. Close price.}
\item{   lag}{ Integer. Number of lag periods.}
\item{   plot}{ Logical. If TRUE plot is returned.}
\item{   ...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


