\name{chaikin}
\alias{chaikin}
\title{Chaikin oscillator}
\description{Compute Chaikin oscillator (Technical Analysis)}

\usage{
chaikin(Close, High = NULL, Low = NULL, 
Vol = NULL, fast.lag = 3, slow.lag = 10, 
plot = TRUE, ...) 
}
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Vol}{ VECTOR. Asset traded Volume.}\item{fast.lag}{ fast.lag}\item{slow.lag}{
slow.lag}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


