\name{chaikin}
\alias{chaikin}

\title{Chaikin oscillator}

\description{Compute Chaikin oscillator (Technical Analysis)}

\usage{chaikin(Close, High = NULL, Low = NULL, Vol = NULL, fast.lag = 3, slow.lag = 10, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{Vol}{ }\item{fast.lag}{ }\item{slow.lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
