\name{chvol}
\alias{Ch.vol}
\title{Chaikin volatility indicator}
\description{Compute Chaikin volatility indicator (Technical Analysis)}
\usage{Ch.vol(High, Low, Close, lag = 5, plot = FALSE, ...) }
\arguments{\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR. Low price.}\item{Close}{ VECTOR.
Close price.}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}

\details{The Chaikin Volatility indicator is the rate of change of the trading range.\cr
The indicator defines volatility as a increasing of the difference between the high and low. A rapid
increases in the
Chaikin Volatility indicate that a bottom is approaching. A slow decrease in the Chaikin Volatility
indicates that a
top is approaching.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
