\name{cleanup}
\alias{cleanup}

\title{Clean memory}

\description{Cleanup environment and (optionally) performs Garbage Collection}

\usage{cleanup(keep = c(), env = parent.frame(), gc = FALSE) }

\arguments{ \item{keep}{ }\item{env}{ }\item{gc}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
