\name{clv}
\alias{clv}

\title{Close Location value oscillator}

\description{Compute Close Location value oscillator (Technical Analysis)}

\usage{clv(Close, High = NULL, Low = NULL, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
