\name{clv}
\alias{clv}
\title{Close Location value oscillator}
\description{Compute Close Location value oscillator (Technical Analysis)}
\usage{clv(Close, High = NULL, Low = NULL, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
