\name{cmf}
\alias{cmf}

\title{Chaikin Money Flow}

\description{Compute Chaikin Money Flow (Technical Analysis)}

\usage{cmf(Close, Low, High, Volume, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{Low}{ }\item{High}{ }\item{Volume}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
