\name{cmof}
\alias{cmof}
\title{Chande Momentum Oscillator}
\description{Compute Chande Momentum Oscillator (Technical Analysis)}
\usage{cmof(X, lag = 5, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


