\name{combine.fs}
\alias{combine.fs}
\title{Combine Multiple Financial Series}
\description{Combine Multiple Financial Series}
\usage{combine.fs(..., which = "Close") }
\arguments{\item{...}{ ...}\item{which}{ which}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
