\name{Decscal}
\alias{Decscal}
\title{Decimal scale}
\description{Compute decimal scale of a vector}
\usage{Decscal(x, scale = 0.1) }
\arguments{\item{x}{ x}\item{scale}{ scale}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
