\name{dgev}
\alias{dgev}
\title{Generalised Extreme Value (GEV)}
\description{Generalised Extreme Value (GEV) - Density function}
\usage{dgev(X, mu = 0, xi = 0.1, sigma = 1) }
\arguments{
\item{X}{ X}
\item{mu}{ mu}
\item{xi}{ xi}
\item{sigma}{ sigma}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


