\name{dgpd}
\alias{dgpd}
\title{Generalised Pareto Distribution (GPD)}
\description{Generalised Pareto Distribution (GPD) - Density function}
\usage{dgpd(X, xi = 0.1, sigma = 1, trsh = 0) }
\arguments{\item{X}{ X}\item{xi}{ xi}\item{sigma}{ sigma}\item{trsh}{ trsh}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


