\name{dpo}
\alias{dpo}

\title{Detrended price oscillator}

\description{Compute Detrended price oscillator (Technical Analysis)}

\usage{dpo(Close, lag = 5, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
