\name{dpo}
\alias{dpo}
\title{Detrended price oscillator}
\description{Compute Detrended price oscillator (Technical Analysis)}
\usage{dpo(Close, lag = 5, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{lag}{ INTEGER. Number of lag
periods.}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


