\name{drawdown}

\alias{drawdown.default}
\alias{drawdown}

\title{Financial Drawdown}

\description{Drawdown risk analysis}

\usage{
drawdown(x, ...) 
\method{drawdown}{default}(x, FUN = max, relative = FALSE, plot = FALSE, ...) 
}

\arguments{ \item{x}{ x}\item{FUN}{ FUN}\item{relative}{ realtive}\item{plot}{ LOGICAL. If TRUE plot
is returned.}\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


