\name{dropn}
\alias{dropn}

\title{Drop N Possible Terms to a Linear Regression Model}

\description{Drop N Possible Terms to a Linear Regression Model}

\usage{dropn(mod, N = 1, ...) }

\arguments{ \item{mod}{ }\item{N}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
