\name{erf}
\alias{erf}
\title{Elder Ray force }
\description{Compute Elder Ray force  (Technical Analysis)}
\usage{erf(Close, High = NULL, Low = NULL, lag = 13, plot = FALSE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


