\name{erfi}
\alias{erfi}
\title{Elder Ray force index}
\description{Compute Elder Ray force index (Technical Analysis)}
\usage{erfi(X, Volume, lag = 13, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{Volume}{ VECTOR. Asset traded Volume.}\item{lag}{ INTEGER. Number of
lag periods.}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or
from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
