\name{erfi}
\alias{erfi}

\title{Elder Ray force index}

\description{Compute Elder Ray force index (Technical Analysis)}

\usage{erfi(X, Volume, lag = 13, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{Volume}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
