\name{fft}

\alias{FFT}
\alias{FFT.default}

\title{Custom Fast Fourier transformation}

\description{Compute Custom Fast Fourier transformation}

\usage{
FFT(x, ...)

\method{FFT}{default}(x, Fs = 1, half = FALSE, 
window = NULL, plot = TRUE, optimised = TRUE, ...) 

}

\arguments{ \item{x}{ x}\item{Fs}{ Fs}\item{half}{ half}\item{window}{ window}
\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{optimised}{ optimised}\item{...}{ Further
arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
