\name{fmlmreg}

\alias{formula.mreg}
\alias{formula.reg}

\title{Extract formula from regression object}

\description{Extract formula from regression ("reg" / "mreg") object}

\usage{
\method{formula}{reg}(x, ...) 
\method{formula}{mreg}(x, ...) 
}

\arguments{ 
\item{x}{ An object of class "reg" / "mreg"}
\item{...}{ Further arguments passed to or from other methods}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


