\name{forcidx}
\alias{forcidx}

\title{Force index}

\description{Compute Force index (Technical Analysis)}

\usage{forcidx(X, Volume, lag = 5, sth = TRUE, sth.lag = 13, mov = sma, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{Volume}{ }\item{lag}{ }\item{sth}{ }\item{sth.lag}{ }\item{mov}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
