\name{FSE.VecAr}
\alias{FSE.VecAr}
\title{VAR Forecast Standard Error }
\description{Compute forecast standard error for VAR model}
\usage{FSE.VecAr(X, steps, ...) }
\arguments{\item{X}{ X}\item{steps}{ steps}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
