\name{fullP}
\alias{fullP}
\title{Full price}
\description{Compute Full price (Technical Analysis)}
\usage{fullP(Close, Open, High, Low, plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Open}{ VECTOR. Open price.}\item{High}{ VECTOR.
High price.}\item{Low}{ VECTOR. Low price.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
