\name{get.fs}
\alias{get.fs}
\title{Download Financial Series data from Yahoo!}
\description{Download Financial Series data from Yahoo!}

\usage{get.fs(symbol = NULL, SName = NULL, 
from = as.Date("1950-01-01"), to = Sys.Date(),
strip.spaces = TRUE, strip.char = ".") }

\arguments{\item{symbol}{ symbol}\item{SName}{ SName}\item{from}{ from}\item{to}{
to}\item{strip.spaces}{ strip.spaces}\item{strip.char}{ strip.char}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
