\name{gev.like}
\alias{gev.like}
\title{GEV - Log Likelihood}
\description{GEV - Log Likelihood}
\usage{gev.like(parms, Xbmax, ...) }
\arguments{\item{parms}{ parms}\item{Xbmax}{ Xbmax}\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
