\name{gev.xi.constraint}
\alias{gev.xi.constraint}
\title{GEV - Domain range for the xi parameter}
\description{GEV - Domain range for the xi parameter}
\usage{gev.xi.constraint(parms, type = c("left", "right", "both"), 
Xbmax, parm.type = c("mu", "VaR","ES"), prob = 0.01, ...) }
\arguments{\item{parms}{ parms}\item{type}{ type}\item{Xbmax}{ Xbmax}\item{parm.type}{
parm.type}\item{prob}{ prob}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
