\name{gini}

\alias{Gini}
\alias{Gini.scorecard}
\alias{Gini.default}

\title{Gini index}

\description{Calculate Gini index based on the results of a classification model.}

\usage{
Gini(x, ...) 
\method{Gini}{default}(x, ...) 
\method{Gini}{scorecard}(x, glob = TRUE, ...) 
}

\arguments{
\item{x}{ An object of class "scorecard"}
\item{glob}{ global}
\item{...}{ Further arguments to or from other methods}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
