\name{Gini}

\alias{Gini}
\alias{Gini.scorecard}

\title{Gini index}

\description{Calculate Gini index based on the results of a classification model.}

\usage{
Gini(x, ...) \cr
Gini.scorecard(sc, glob = TRUE, ...) 
}

\arguments{
\item{x}{ }
\item{sc}{ }
\item{glob}{ }
\item{...}{ }
}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
