\name{GKgamma}

\alias{GKgamma}

\title{GK Gamma}

\description{GK Gamma statistic}

\usage{GKgamma(target, pred, ...) }

\arguments{\item{target}{ }\item{pred}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
