\name{gpd.sigma.constraint}
\alias{gpd.sigma.constraint}
\title{GPD - Domain range for the sigma parameter}
\description{GPD - Domain range for the sigma parameter}

\usage{gpd.sigma.constraint(parms, type = c("left", "right", "both"), 
Xtail, trsh = 0, ...) }
\arguments{\item{parms}{ parms}\item{type}{ type}\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{...}{
Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
