\name{gpd.VaR.contour}
\alias{gpd.VaR.contour}

\title{GPD - VaR Joint Confidence Intervals by Profile Likelihood}

\description{}

\usage{gpd.VaR.contour(Xtail, trsh = 0, VaR = trsh + 10^-5, xi = 0.1, alpha = 0.01, df = 2, N, prob
= alpha[1]) }

\arguments{ \item{Xtail}{ }\item{trsh}{ }\item{VaR}{ }\item{xi}{ }\item{alpha}{ }\item{df}{
}\item{N}{ }\item{prob}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
