\name{gpdboot}
\alias{gpdboot}
\title{GPD - parameters bootstrapping}
\description{GPD - parameters bootstrapping}
\usage{gpdboot(Xtail, trsh = 0, xi = NULL, sigma = NULL, nboots = 100, ...) }
\arguments{\item{Xtail}{ Xtail}\item{trsh}{ trsh}\item{xi}{ xi}\item{sigma}{ sigma}\item{nboots}{
nboots}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
