\name{grad}
\alias{grad}
\title{Compute numerical gradient of a function}
\description{Plotting tools}
\usage{grad(func = NULL, x, scalar = TRUE, eps = sqrt(.Machine$double.neg.eps), ...) }
\arguments{\item{func}{ func}\item{x}{ x}\item{scalar}{ scalar}\item{eps}{ eps}\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
