\name{GrangCas}

\alias{GrangCas.VecAr}

\title{Granger Causality test}

\description{Perform Granger causality test for parameters of VAR model}

\usage{
\method{GrangCas}{VecAr}(X, cause = NULL, ...) 
}

\arguments{ \item{X}{ X}\item{cause}{ cause}\item{...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
