\name{Graph_utils}
\alias{getTheme}
\alias{getCurrentTheme}
\alias{getThemeAttr}
\alias{setCurrentTheme}
\alias{setThemeAttr}
\alias{transition}
\alias{gradient}
\alias{jet.colors}
\alias{loadThemes}
\alias{override.list}
\alias{set.bg}
\alias{set.bg3d}
\alias{create.empty.plot}
\alias{optimize.polycords}
\alias{shade.plot} 
\alias{comma.Fmt} 
\alias{comma.kFmt} 
\alias{comma.mFmt}
\alias{SI.format} 
\alias{apply.format} 
\alias{get.plot.layout}
\alias{get.plot.params}
\alias{getProjectionMatrix}
\alias{setProjectionMatrix}
\title{ RAdamant Graphical utilities}
\description{Graphical utilities used by the plotting functions}
\author{ RAdamant Development Team \email{team@r-adamant.org} }
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
