\name{heas}
\alias{he_as}
\title{Heikin - Ashi techniques}
\description{Compute Heikin - Ashi techniques (Technical Analysis)}
\usage{he_as(Close, Open, High, Low, plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Open}{ VECTOR. Open price.}\item{High}{ VECTOR.
High price.}\item{Low}{ VECTOR. Low price.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}

\details{A type of candlestick chart that shares many characteristics with standard candlestick
charts, but differs because of the values used to create each bar.\cr
The Heikin-Ashi technique is used by technical traders to identify a given trend more easily. Hollow
candles with no lower shadows are used to signal a strong uptrend, while filled candles with no
higher shadow are used to identify a strong downtrend.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
