\name{hhv}
\alias{hhv}

\title{Highest high}

\description{Compute Highest high (Technical Analysis)}

\usage{hhv(X, lag, na.rm = TRUE) }

\arguments{ \item{X}{ }\item{lag}{ }\item{na.rm}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
