\name{hill}

\alias{Hill}

\title{Hill function}

\description{Hill function: Approximated gamma parameter of the Generalised Pareto distribution}

\usage{
Hill(X, trsh)
}

\arguments{
\item{X}{ Input matrix/sequence. Sequences are treated as one column matrices.}
\item{trsh}{ vector of probability threshold (interval [0, 1])}
}

\value{ A matrix length(trsh) by NCOL(X) of computed quantiles}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
