\name{hroi}
\alias{hroi}
\title{Historical Returns on Investments}
\description{Historical Returns on Investments}

\usage{hroi(X, lag = 1, mode = c("auto", "range", "selected"), 
autolag.start = 1, range.step = 1, log = TRUE, 
VaR.type = "norm", p = 0.05, ...) }

\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{mode}{
mode}\item{autolag.start}{ autolag.start}\item{range.step}{ range.step}\item{log}{
log}\item{VaR.type}{ VaR.type}\item{p}{ p}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
