\name{ichkh}
\alias{Ichkh}
\title{Ichimoku Kinko Hyo}
\description{Compute Ichimoku Kinko Hyo (Technical Analysis)}
\usage{Ichkh(Close, High, Low, plot = FALSE, ...) }
\arguments{ \item{Close}{ close}\item{High}{ high}\item{Low}{ low}\item{plot}{ LOGICAL. If TRUE plot
is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


