\name{impulse}
\alias{impulse}
\title{Unitary impulse}
\description{Generates an impulse sequence of specified length}
\usage{impulse(N, value = 1) }
\arguments{ 
\item{N}{ Length of the impulse}
\item{value}{ value of the impulse (Default = 1)}}
\value{Impulse sequence of specified length}
\author{ RAdamant Development Team \email{team@r-adamant.org} }
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
