\name{in2woe}
\alias{input2woe}
\title{Data to Weight of Evidence}
\description{Transform input data according to weight of evidence}
\usage{input2woe(data, nseg, woe, na.replace, 
seg.type = c("freq_equal", "width_equal")) }
\arguments{\item{data}{ data}\item{nseg}{ nseg}\item{woe}{ woe}\item{na.replace}{
na.replace}\item{seg.type}{ seg.type}}
\author{RAdamant Development Team \email{team@r-adamant.org} }
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
