\name{Inertia}
\alias{Inertia}

\title{Inertia oscillator}

\description{Compute Inertia oscillator (Technical Analysis)}

\usage{Inertia(X, lag, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
