\name{IRS.VecAr}
\alias{IRS.VecAr}
\title{VAR Impulse response}
\description{Compute and plot Impulse response function calculated for VAR model}
\usage{IRS.VecAr(X, imp, resp = NULL, steps = 5, cum = TRUE, ortho = FALSE, ...) }
\arguments{\item{X}{ X}\item{imp}{ imp}\item{resp}{ resp}\item{steps}{ steps}\item{cum}{
cum}\item{ortho}{ ortho}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
