\name{JB.test}
\alias{JB.test}

\title{Jaques-Brera normality test}

\description{Compute Jaques-Brera normality test for each column of X}

\usage{
JB.test(X, plot=FALSE) 
}

\arguments{ 
\item{X}{ Matrix of data series (one column per variable)}
\item{plot}{ LOGICAL. Return histogram.}
}

\details{}

\value{Matrix of Jaques-Brera scores and P-Value}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{ \code{\link{kurt}}, \code{\link{skew}}}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
