\name{kelt}
\alias{kelt}

\title{Keltner channel}

\description{Compute Keltner channel (Technical Analysis)}

\usage{kelt(Close, High, Low, mult = 2, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{mult}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
