\name{kri}
\alias{kri}

\title{Kairi Relative Index}

\description{Compute Kairi Relative Index (Technical Analysis)}

\usage{kri(X, lag1 = 10, lag2 = 20, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag1}{ }\item{lag2}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
