\name{kvo}
\alias{kvo}
\title{Klinger oscillator}
\description{Compute Klinger oscillator (Technical Analysis)}

\usage{kvo(Close, High = NULL, Low = NULL, 
Vol = NULL, cumulative = FALSE, plot = TRUE, ...) }

\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Vol}{ VECTOR. Asset traded Volume.}\item{cumulative}{ cumulative}\item{plot}{
LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}

\details{The Klinger Oscillator is a volume- and price-based oscillator intended to measure both
short- and long-term money flows into and out of a security.\cr
The Klinger Volume Oscillator measures trends of money flows based on volume.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
