\name{kvo}
\alias{kvo}
\title{Klinger oscillator}
\description{Compute Klinger oscillator (Technical Analysis)}

\usage{kvo(Close, High = NULL, Low = NULL, 
Vol = NULL, cumulative = FALSE, plot = TRUE, ...) }

\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{Vol}{Vector. Asset traded Volume.}
\item{cumulative}{ cumulative}
\item{plot}{
LOGICAL. If TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods.}}




\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


