\name{liftgain}

\alias{Gain}
\alias{Gain.scorecard}
\alias{Lift}
\alias{Lift.scorecard}

\title{Lift and Gain plot}

\description{Plot cumulative Gain and Lift chart for a classification model}

\usage{
Gain(x, ...) 
Lift(x, ...) 
\method{Gain}{scorecard}(x, pc = 0.1, ...) 
\method{Lift}{scorecard}(x, pc = 0.1, ...) 
}

\arguments{
\item{x}{ x}
\item{pc}{ pc}
\item{...}{ Further arguments to or from other methods}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
