\name{lkmgarch}

\alias{like.mgarch}

\title{MGARCH likelihood function}

\description{Calculate MGARCH likelihood function}

\usage{like.mgarch(theta, x, Y, order, prob=c("norm","ged","t")) }

\arguments{
\item{theta}{ theta}
\item{x}{ x}
\item{Y}{ Y}
\item{order}{ order}
\item{prob}{ prob}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


