\name{llv}
\alias{llv}
\title{Lowest low}
\description{Compute Lowest low (Technical Analysis)}
\usage{llv(X, lag, na.rm = TRUE) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{na.rm}{ na.rm}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


