\name{Logger}
\alias{Logger}

\title{Main logging function}

\description{Create Log for the functions contained in the package}

\usage{Logger(message = "", from = deparse(sys.call(sys.parent())), level = 1, line = NA, env =
getOption("RAdamant"), console = getConsoleLogging(env = env), logfile = getLogFile(env = env)) }

\arguments{ \item{message}{ }\item{from}{ }\item{level}{ }\item{line}{ }\item{env}{ }\item{console}{ }\item{logfile}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
