\name{macd}
\alias{macd}
\title{Moving Average Convergence / Divergence}
\description{Compute Moving Average Convergence / Divergence (Technical Analysis)}
\usage{macd(X, fast.lag = 12, slow.lag = 26, signal.lag = 14, plot = TRUE, ...) }
\arguments{
\item{X}{ X}
\item{fast.lag}{ fast.lag}
\item{slow.lag}{ slow.lag}
\item{signal.lag}{
signal.lag}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


