\name{mass}
\alias{mass}
\title{Mass indicator}
\description{Compute Mass indicator (Technical Analysis)}
\usage{mass(High, Low, Close = NULL, lag = 9, plot = FALSE, ...) }
\arguments{\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR. Low price.}\item{Close}{ VECTOR.
Close price.}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


