\name{mass}
\alias{mass}

\title{Mass indicator}

\description{Compute Mass indicator (Technical Analysis)}

\usage{mass(High, Low, Close = NULL, lag = 9, plot = FALSE, ...) }

\arguments{ \item{High}{ }\item{Low}{ }\item{Close}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
