\name{mcgind}
\alias{mcgind}

\title{ McGinley Dynamic Indicator}

\description{Compute  McGinley Dynamic Indicator (Technical Analysis)}

\usage{mcgind(X, lag = 12, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
