\name{mcosc}
\alias{mcosc}
\title{McClellan Oscillator}
\description{Compute McClellan Oscillator (Technical Analysis)}
\usage{mcosc(X, fast.lag = 19, slow.lag = 39, hist.lag = 9, plot = TRUE, ...) }
\arguments{
\item{X}{ X}
\item{fast.lag}{ fast.lag}
\item{slow.lag}{ slow.lag}
\item{hist.lag}{
hist.lag}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


