\name{mcplot}
\alias{mcplot}
\title{Muliple correlation plot}
\description{Multiple correlation plot}

\usage{
mcplot(X
	, hist.nclass = 10
	, theme.params = getCurrentTheme()
	, coLin = TRUE
	, main = ifelse(coLin, "Co-Linearity Analysis", "Multi-Correlation Analysis")
	, new.device = FALSE
	, ...
	) 
}

\arguments{
\item{X}{ X}
\item{hist.nclass}{ hist.nclass}
\item{theme.params}{theme.params}
\item{coLin}{ coLin}
\item{main}{ main}
\item{new.device}{ new.device}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


