\name{mcsi}
\alias{mcsi}
\title{McClellan Summation Index}
\description{Compute McClellan Summation Index (Technical Analysis)}
\usage{mcsi(matr, nr, nc, lag1, lag2, plot = FALSE, ...) }
\arguments{\item{matr}{ matr}\item{nr}{ nr}\item{nc}{ nc}\item{lag1}{ lag1}\item{lag2}{
lag2}\item{plot}{ LOGICAL. If
TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
