\name{mflow}
\alias{Mflow}
\title{Money flow }
\description{Compute Money flow  (Technical Analysis)}
\usage{Mflow(Close, High, Low, Volume, plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Volume}{ VECTOR. Asset traded Volume.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


