\name{Minmaxscal}
\alias{Minmaxscal}

\title{Mini/Max Scale}

\description{Compute minimum / maximum scale of a vector}

\usage{Minmaxscal(x, tmin = 0, tmax = 1) }

\arguments{ \item{x}{ }\item{tmin}{ }\item{tmax}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
