\name{mom}
\alias{mom}

\title{Momentum oscillator}

\description{Compute Momentum oscillator (Technical Analysis)}

\usage{mom(X, lag = 5, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
