\name{Moments}
\alias{moments}

\title{Main Moments}

\description{Calculate sample moments on each columns of X}

\usage{moments(X) }

\arguments{ 
\item{X}{ Matrix of data series (one column per variable)}
}

\details{}

\value{Matrix of moments}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org} }

\note{}

\seealso{\code{\link{JB.test}} , \code{\link{skew}} , \code{\link{kurt}} }

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
