\name{moments}
\alias{moments}
\title{Main Moments}
\description{Calculate sample moments on each columns of X}
\usage{moments(X) }
\arguments{ 
\item{X}{ Matrix of data series (one column per variable)}
}
\value{Matrix of moments}
\author{RAdamant Development Team \email{team@r-adamant.org} }
\seealso{\code{\link{JB.test}} , \code{\link{skew}} , \code{\link{kurt}} }
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
