\name{mqt}

\alias{mqt}

\title{Multiple t quantile}

\description{Compute quantiles from Students T distribution for multiple degrees of freedom values}

\usage{
mqt(p, df, ...)
}

\arguments{
\item{p}{ Vector of probabilities (DEFAULT = 0.05)}
\item{df}{ Vector of degrees of freedom}
\item{...}{ Further arguments to and from other methods}
}

\value{A matrix length(p) by length(df) of computed quantiles}

\author{RAdamant Development Team \email{team@r-adamant.org}}




