\name{mreg}
\alias{mreg}

\title{Multiple regressions}

\description{Multiple regressions}

\usage{mreg(Y, X, xlabels = NULL, tick.step = 1, backtest = 0, stress.idx = c(), type = "simple",
model = "lm", ci = 0.95, max.vars = NCOL(X), intercept = TRUE, family = gaussian, weights = NULL,
plot = TRUE, scope = NULL, trace = FALSE, ...) }

\arguments{ \item{Y}{ }\item{X}{ }\item{xlabels}{ }\item{tick.step}{ }\item{backtest}{ }\item{stress.idx}{ }\item{type}{ }\item{model}{ }\item{ci}{ }\item{max.vars}{ }\item{intercept}{ }\item{family}{ }\item{weights}{ }\item{plot}{ }\item{scope}{ }\item{trace}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
