\name{newsimp}

\alias{newsimp.default}
\alias{newsimp.Garch}
\alias{newsimp}

\title{News impact curve }

\description{Compute News impact curve }

\usage{
newsimp(x, ...)
\method{newsimp}{Garch}(x, plot = TRUE, ...)
\method{newsimp}{default}(x, theta, order, 
type=c("garch","mgarch", "egarch","tgarch"), 
plot=FALSE, ...) 
}

\arguments{ 
\item{x}{ x}
\item{theta}{ theta}
\item{order}{ order}
\item{type}{ type}
\item{plot}{ LOGICAL. If TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
