\name{normfit}
\alias{norm.fit}
\title{Fit normal distribution}
\description{Fit normal distribution}
\usage{norm.fit(x, n = 200, range = NULL, ...) }
\arguments{\item{x}{ x}\item{n}{ n}\item{range}{ range}\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
