\name{Obv}
\alias{Obv}
\title{On Balance Volume oscillator}
\description{Compute On Balance Volume oscillator (Technical Analysis)}
\usage{Obv(Close, Volume) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Volume}{ VECTOR. Asset traded Volume.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
