\name{Obv}
\alias{Obv}

\title{On Balance Volume oscillator}

\description{Compute On Balance Volume oscillator (Technical Analysis)}

\usage{Obv(Close, Volume) }

\arguments{ \item{Close}{ }\item{Volume}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
