\name{Oscil}

\alias{oscil.default}
\alias{oscil}

\title{Oscillator default method}

\description{Compute Oscillator (Technical Analysis)}

\usage{
oscil(X, ...)
\method{oscil}{default}(X, Y, pc = FALSE, type = "oscil", ...) 
}


\arguments{ \item{X}{ X}\item{Y}{ Y}\item{pc}{ pc}\item{type}{ type}\item{...}{ Further arguments to
or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
