\name{pchan}
\alias{Pchan}
\title{Price channel}
\description{Compute Price channel (Technical Analysis)}
\usage{
Pchan(CLose
     , High
     , Low
     , lag = 20
     , na.rm = TRUE
     , plot = FALSE
     , ...
     ) 
}

\arguments{
\item{CLose}{ CLose}
\item{High}{Vector. High price.}
\item{Low}{Vector. Low price.}
\item{lag}{Integer. Number of lag periods.}
\item{na.rm}{ na.rm}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods.}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


