\name{Pchan}
\alias{Pchan}
\title{Price channel}
\description{Compute Price channel (Technical Analysis)}
\usage{Pchan(CLose, High, Low, lag = 20, na.rm = TRUE, plot = FALSE, ...) }
\arguments{\item{CLose}{ CLose}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR. Low
price.}\item{lag}{ INTEGER. Number of lag periods.}\item{na.rm}{ na.rm}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
