\name{pdfhit}
\alias{PDFHit}
\title{Density of Hitting probability}
\description{Density for the First Hitting time}
\usage{PDFHit(t, B = 0, S0 = 0, mi, sigma, cumul = FALSE, plot = FALSE, ...) }
\arguments{\item{t}{ t}\item{B}{ B}\item{S0}{ S0}\item{mi}{ mi}\item{sigma}{ sigma}\item{cumul}{
cumul}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


