\name{perf}
\alias{Perf}
\title{Performance indicator}
\description{Compute Performance indicator (Technical Analysis)}
\usage{Perf(X, ini.per = 1, cut = TRUE, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{ini.per}{ ini.per}
\item{cut}{ cut}
\item{plot}{Logical. If TRUE plot is
returned.}
\item{...}{ Further arguments to or from other methods.}}




\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


