\name{perf}
\alias{Perf}
\title{Performance indicator}
\description{Compute Performance indicator (Technical Analysis)}
\usage{Perf(X, ini.per = 1, cut = TRUE, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{ini.per}{ ini.per}\item{cut}{ cut}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}

\details{The Performance indicator displays the percentage difference between the price today and
the price at the start of the data series. It is also known as a normalized price. It can be useful
for comparing the performance of two securities or a security and an index. }



\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
