\name{pfe}
\alias{pfe}
\title{Polarized fractal efficiency}
\description{Compute Polarized fractal efficiency (Technical Analysis)}
\usage{pfe(X, lag = 9, corr_fact = 200, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{corr_fact}{
corr_fact}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
