\name{pfe}
\alias{pfe}

\title{Polarized fractal efficiency}

\description{Compute Polarized fractal efficiency (Technical Analysis)}

\usage{pfe(X, lag = 9, corr_fact = 200, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{corr_fact}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
