\name{pgev}
\alias{pgev}
\title{Generalised Extreme Value (GEV) - Probability function}
\description{Generalised Extreme Value (GEV) - Probability function}
\usage{pgev(X, mu = 0, xi = 0.1, sigma = 1) }
\arguments{\item{X}{ X}\item{mu}{ mu}\item{xi}{ xi}\item{sigma}{ sigma}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
