\name{pgpd}
\alias{pgpd}
\title{Generalised Pareto Distribution (GPD) - Probability function}
\description{Generalised Pareto Distribution (GPD) - Probability function}
\usage{pgpd(Q, xi = 0.1, sigma = 1, trsh = 0) }
\arguments{
\item{Q}{ Q}
\item{xi}{ xi}
\item{sigma}{ sigma}
\item{trsh}{ trsh}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


