\name{PHI.VecAr}
\alias{PHI.VecAr}
\title{VAR - PHI}
\description{Estimate PHI matrix for MA (Wold) representation of VAR model}
\usage{PHI.VecAr(X, steps, ortho = FALSE, ...) }
\arguments{ \item{X}{ OBJECT of class "VecAR"}\item{steps}{ INTEGER. Number of steps
ahead.}\item{ortho}{ LOGICAL. If TRUE matrix is orthogonal}\item{...}{ Further arguments to or from
other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
