\name{plot.sme}
\alias{plot.sme}
\title{Plot Sample Mean Excess class}
\description{Plotting function for Sample Mean Excess class}
\usage{
\method{plot}{sme}(x, main = attr(x, "desc"), xtitle = get.col.names(attr(x, "data")), ...) }
\arguments{ \item{x}{ OBJECT of class "sme".}\item{main}{ main}\item{xtitle}{ xtitle}
\item{...}{ Further arguments to or from other methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
