\name{plotmreg}
\alias{plot.mreg}
\title{Plot function for mreg}
\description{Plot function for class 'mreg'}
\usage{
\method{plot}{mreg}(x, ...) }
\arguments{ \item{x}{ OBJECT of class "mreg".}\item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
