\name{pmreg}
\alias{print.mreg}
\title{Print function for mreg}
\description{Print function for class 'mreg'}
\usage{
\method{print}{mreg}(x, ...) 
}
\arguments{ \item{x}{ OBJECT of class "mreg".} \item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


