\name{ppo}
\alias{ppo}
\title{Percentage Price oscillator}
\description{Compute Percentage Price oscillator (Technical Analysis)}
\usage{ppo(X, fast.lag = 10, slow.lag = 30, plot = TRUE, ...) }
\arguments{\item{X}{ X}\item{fast.lag}{ fast.lag}\item{slow.lag}{ slow.lag}\item{plot}{ LOGICAL. If
TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
