\name{prbsar}
\alias{prbsar}
\title{Parabolic Stop and Reverse (PSAR)}
\description{Compute Parabolic Stop and Reverse (PSAR) (Technical Analysis)}
\usage{prbsar(Close, High, Low, accel = c(0.02, 0.2), plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{accel}{ accel}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
