\name{prbsar}
\alias{prbsar}
\title{Parabolic Stop and Reverse (PSAR)}
\description{Compute Parabolic Stop and Reverse (PSAR) (Technical Analysis)}
\usage{
prbsar(Close, High, Low
    , accel = c(0.02, 0.2)
    , plot = FALSE
    , ...
    )
}
\arguments{
\item{ Close}{Vector. Close price.}
\item{ High}{Vector. High price.}
\item{ Low}{Vector. Low price.}
\item{ accel}{ accel}
\item{ plot}{Logical. If TRUE plot is returned.}
\item{ ...}{ Further arguments to or from other methods.}
}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


