\name{Pred.Garch}

\alias{predict.Garch}

\title{Predict Garch model}

\description{Predict Garch model}

\usage{
\method{predict}{Garch}(object, plot = TRUE, ...) 
}

\arguments{ \item{object}{ OBJECT of class "Garch".}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
