\name{preder}
\alias{pred_error}
\title{Prediction error}
\description{Calculate Prediction error for model evaluation }
\usage{pred_error(actual, predicted, pc = FALSE) }
\arguments{\item{actual}{ actual}\item{predicted}{ predicted}\item{pc}{ pc}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
