\name{predmreg}
\alias{predict.mreg}
\title{Predict method for Multiple regressions}
\description{Predict function for class 'mreg'}
\usage{
\method{predict}{mreg}(object, ...) 
}
\arguments{ \item{object}{ OBJECT of class "mreg".}\item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
