\name{pro}
\alias{pro}

\title{Price oscillator}

\description{Compute Price oscillator (Technical Analysis)}

\usage{pro(Close, fast.lag = 5, slow.lag = 10, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{fast.lag}{ }\item{slow.lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
