\name{probhit}
\alias{ProbHit}
\title{Probability of first hit (Brownian motion)}
\description{Calcualte probability to Hit a barrier}
\usage{ProbHit(B = 0, S0 = 0, mi, sigma) }
\arguments{\item{B}{ B}\item{S0}{ S0}\item{mi}{ mi}\item{sigma}{ sigma}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


