\name{PtfBeta}
\alias{PtfBeta}

\title{Portfolio Beta}

\description{Get portfolio Beta}

\usage{PtfBeta(beta, w = NULL, glob = TRUE) }

\arguments{ \item{beta}{ }\item{w}{ }\item{glob}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
