\name{PtfOpt}
\alias{PtfOpt}

\title{Mean-Variance optimum portfolio}

\description{Calculate mean-variance efficient portfolio}

\usage{PtfOpt(ret = NULL, ptf = NULL, mi = NULL, SIGMA = NULL, volatility = TRUE, ...) }

\arguments{ \item{ret}{ }\item{ptf}{ }\item{mi}{ }\item{SIGMA}{ }\item{volatility}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
