\name{PtfRet}
\alias{PtfRet}
\title{Portfolio returns}
\description{Calculate portfolio returns}
\usage{PtfRet(PTF, w = NULL, glob = TRUE, calc.ret = FALSE, ...) }
\arguments{\item{PTF}{ PTF}\item{w}{ w}\item{glob}{ glob}\item{calc.ret}{ calc.ret}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
