\name{ptfutil}
\alias{PtfUtility}
\title{Portfolio Utility}
\description{Calculate utility and plot for efficient portfolio}

\usage{PtfUtility(PTF = NULL, W, R = NULL, SIGMA = NULL, 
af = 3, plot = TRUE, plot.mv = FALSE,
plot.mu = FALSE, ...) }

\arguments{\item{PTF}{ PTF}\item{W}{ W}\item{R}{ R}\item{SIGMA}{ SIGMA}\item{af}{ af}\item{plot}{
LOGICAL. If TRUE plot is returned.}\item{plot.mv}{ plot.mv}\item{plot.mu}{ plot.mu}\item{...}{
Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
