\name{PtfVar}
\alias{PtfVar}

\title{Portfolio variance}

\description{Calculate portfolio variance}

\usage{PtfVar(PTF, w = NULL, glob = TRUE, vol = FALSE, calc.ret = FALSE, ...) }

\arguments{ \item{PTF}{ }\item{w}{ }\item{glob}{ }\item{vol}{ }\item{calc.ret}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
