\name{pvt}
\alias{pvt}

\title{Price Volume trend indicator}

\description{Compute Price Volume trend indicator (Technical Analysis)}

\usage{pvt(Close, Volume, lag = 5, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{Volume}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
