\name{pvt}
\alias{pvt}
\title{Price Volume trend indicator}
\description{Compute Price Volume trend indicator (Technical Analysis)}
\usage{pvt(Close, Volume, lag = 5, plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{Volume}{ VECTOR. Asset traded
Volume.}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}

\details{Price Volume Trend combines percentage price change and volume to confirm the strength of
price trends or through divergences, warn of weak price moves. Unlike other price-volume indicators,
the Price Volume Trend takes into consideration the percentage increase or decrease in price, rather
than just simply adding or subtracting volume based on whether the current price is higher than the
previous days price.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
