\name{qgev}
\alias{qgev}
\title{Generalised Extreme Value (GEV) - Quantile function}
\description{Generalised Extreme Value (GEV) - Quantile function}
\usage{qgev(P, mu = 0, xi = 0.1, sigma = 1) }
\arguments{\item{P}{ P}\item{mu}{ mu}\item{xi}{ xi}\item{sigma}{ sigma}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


