\name{qgpd}
\alias{qgpd}
\title{Generalised Pareto Distribution (GPD) - Quantile function}
\description{Generalised Pareto Distribution (GPD) - Quantile function}
\usage{qgpd(P, xi = 0.1, sigma = 1, trsh = 0) }
\arguments{\item{P}{ P}\item{xi}{ xi}\item{sigma}{ sigma}\item{trsh}{ trsh}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


