\name{relvol}
\alias{RelVol}
\title{Relative Volatility oscillator}
\description{Compute Relative Volatility oscillator (Technical Analysis)}
\usage{RelVol(Close, sdlag = 9, lag = 5) }
\arguments{
\item{Close}{Vector. Close price.}
\item{sdlag}{ sdlag}
\item{lag}{Integer. Number of lag
periods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


