\name{RelVol}
\alias{RelVol}
\title{Relative Volatility oscillator}
\description{Compute Relative Volatility oscillator (Technical Analysis)}
\usage{RelVol(Close, sdlag = 9, lag = 5) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{sdlag}{ sdlag}\item{lag}{ INTEGER. Number of lag
periods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
