\name{rgpd}
\alias{rgpd}
\title{Generalised Pareto Distribution (GPD) - Random Numbers Generator}
\description{Generalised Pareto Distribution (GPD) - Random Numbers Generator}
\usage{rgpd(n, xi = 0.1, sigma = 1, trsh = 0) }
\arguments{ \item{n}{ n}\item{xi}{ xi}\item{sigma}{ sigma}\item{trsh}{ trsh}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
