\name{roc}
\alias{roc}

\title{Rate of Change index}

\description{Compute Rate of Change index (Technical Analysis)}

\usage{roc(X, lag = 5, pc = TRUE, plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{pc}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
