\name{roc}
\alias{roc}
\title{Rate of Change index}
\description{Compute Rate of Change index (Technical Analysis)}
\usage{roc(X, lag = 5, pc = TRUE, plot = TRUE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{pc}{ pc}\item{plot}{
LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


