\name{ROCplot}
\alias{ROCplot}
\alias{ROCplot.scorecard}
\title{ROC analysis}
\description{Plot ROC curve}

\usage{
ROCplot(x, ...) 
\method{ROCplot}{scorecard}(x, ...) 
}

\arguments{
\item{x}{ x}
\item{...}{ Further arguments to or from other methods}
}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
