\name{rsi}
\alias{rsi}

\title{Relative strength indicator}

\description{Compute Relative strength indicator (Technical Analysis)}

\usage{rsi(X, lag, plot = FALSE, ...) }

\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods.}
}



\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


