\name{rsi}
\alias{rsi}

\title{Relative strength indicator}

\description{Compute Relative strength indicator (Technical Analysis)}

\usage{rsi(X, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{plot}{ }\item{...}{ }}

\details{

}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
