\name{rvi}
\alias{rvi}
\title{Relative Vigor indicator}
\description{Compute Relative Vigor indicator (Technical Analysis)}
\usage{rvi(Close, High = NULL, Low = NULL, Open = NULL, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Open}{ VECTOR. Open price.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}

\details{An indicator used in technical analysis that measures the conviction of a recent price
action and the likelihood that it will continue. The RVI compares the positioning of a security
closing price relative to its price range, and the result is smoothed by calculating an exponential
moving average of the values.}


\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
