\name{rvi}
\alias{rvi}

\title{Relative Vigor indicator}

\description{Compute Relative Vigor indicator (Technical Analysis)}

\usage{rvi(Close, High = NULL, Low = NULL, Open = NULL, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{Open}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
