\name{rvi}
\alias{rvi}
\title{Relative Vigor indicator}
\description{Compute Relative Vigor indicator (Technical Analysis)}
\usage{rvi(Close, High = NULL, Low = NULL, Open = NULL, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Open}{ VECTOR. Open price.}\item{plot}{ LOGICAL. If TRUE plot is
returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
