\name{sampmom}
\alias{SampMom}
\title{Sample moments (Brownian motion)}
\description{Calculate sample moments of a Brownian motion}
\usage{SampMom(P, X, moms = 1:2) }
\arguments{\item{P}{ P}\item{X}{ X}\item{moms}{ moms}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


