\name{sensan}

\alias{sensAnalysis.default}
\alias{sensAnalysis}

\title{Sensitivity analysis default method}

\description{Sensitivity analysis default method}

\usage{
sensAnalysis(X, ...)
\method{sensAnalysis}{default}(X, win.size = length(coef(X)), plot = FALSE, ...) 
}

\arguments{ \item{X}{ X}\item{win.size}{ win.size}\item{plot}{ LOGICAL. If TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
