\name{sensanrg}
\alias{sensAnalysis.reg}
\title{Sensitivity analysis method for reg}


\description{Sensitivity analysis method for reg}

\usage{
\method{sensAnalysis}{reg}(X, ...) 
}

\arguments{ \item{X}{ OBJECT of class "reg".}\item{...}{ Further arguments to or from other
methods}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


