\name{sme}
\alias{sme}
\title{Sample Mean Excess function}
\description{Sample Mean Excess function}
\usage{sme(X, plot = TRUE, ...) }
\arguments{\item{X}{ X}\item{plot}{ plot}\item{...}{ ...}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
