\name{somerd}
\alias{SomerD}
\title{Somers D}
\description{Calculate Somers d statistic}
\usage{SomerD(target, pred, ...) }
\arguments{\item{target}{ target}\item{pred}{ pred}\item{...}{ Further arguments to or from other
methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
