\name{specgram}
\alias{specgram}

\title{Spectrogram}

\description{Create spectrogram for FFT}

\usage{specgram(X, win.size = max(1, NROW(X)/20), plot = TRUE, ...) }

\arguments{ \item{X}{ }\item{win.size}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
