\name{splitWindow}
\alias{splitWindow}

\title{Sliding windows}

\description{}

\usage{splitWindow(N, direction = c("forward", "backward"), mode = c("EW", "SW"), from = NULL, win.size = 1, by = 1, labels = 1:N, ...) }

\arguments{ \item{N}{ }\item{direction}{ }\item{mode}{ }\item{from}{ }\item{win.size}{ }\item{by}{ }\item{labels}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
