\name{starc}
\alias{starc}

\title{Stoller Starc bands}

\description{Compute Stoller Starc bands (Technical Analysis)}

\usage{starc(Close, High = NULL, Low = NULL, atr.mult = 2, lag = 5, atr.lag =
14, mov = c("sma", "ema", "wma"), plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{atr.mult}{ }\item{lag}{ }\item{atr.lag}{ }\item{mov}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
