\name{starc}
\alias{starc}
\title{Stoller Starc bands}
\description{Compute Stoller Starc bands (Technical Analysis)}
\usage{starc(Close, High = NULL, Low = NULL, atr.mult = 2, lag = 5, atr.lag =
14, mov = c("sma", "ema", "wma"), plot = FALSE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{atr.mult}{ atr.mult}\item{lag}{ INTEGER. Number of lag periods.}\item{atr.lag}{
atr.lag}\item{mov}{ mov}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further
arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
