\name{statbar}
\alias{statusbar}
\title{Status bar}

\description{Interactive status bar for console logging}

\usage{statusbar(message = "Computing..", status = 0, n = 1, N = 1, step = 0.01) }

\arguments{\item{message}{ message}\item{status}{ status}\item{n}{ n}\item{N}{ N}\item{step}{ step}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
