\name{stepmat}
\alias{StepMat}
\title{Step matrix for binomial tree}
\description{Simulate binomial path of a binomial tree}
\usage{StepMat(init, n_step, up, down) }
\arguments{\item{init}{ init}\item{n_step}{ n_step}\item{up}{ up}\item{down}{ down}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


