\name{strvar}

\alias{Strvar.VecAr}

\title{Structural Vector Autoregressive model}

\description{Estimate Structural Vector Autoregressive model}

\usage{
Strvar.VecAr(X, A = NULL, B = NULL, inter = FALSE, ...) 
}

\arguments{ \item{X}{ X}\item{A}{ A}\item{B}{ B}\item{inter}{ inter}\item{...}{ Further arguments to
or from other methods}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


