\name{Styles}
\alias{Styles}

\title{Styles analysis (portfolio)}

\description{Perform Style analysis  }

\usage{Styles(FUND, IND, W, lower = NULL, upper = NULL, ...) }

\arguments{ \item{FUND}{ }\item{IND}{ }\item{W}{ }\item{lower}{ }\item{upper}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
