\name{SummaryDD}
\alias{SummaryDD}
\title{Summary drawdown}
\description{Summary function for drawdown}
\usage{SummaryDD(DD) }

\arguments{ \item{DD}{ OBJECT of class "drawdown"}}

\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
