\name{Swing}
\alias{Swing}

\title{Swing Index}

\description{Calculate Swing index (Technical Analysis)}

\usage{Swing(Close, High, Low, Open, ret_cum = FALSE, plot = FALSE)}

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{Open}{ }\item{plot}{ }\item{ret_cum}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
