\name{swing}
\alias{Swing}
\title{Swing Index}
\description{Calculate Swing index (Technical Analysis)}
\usage{Swing(Close, High, Low, Open, ret_cum = FALSE, plot = FALSE, ...)}
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{Open}{ VECTOR. Open price.}\item{ret_cum}{ ret_cum}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


