\name{thigh}
\alias{thigh}
\title{True High oscillator}
\description{Compute True High oscillator (Technical Analysis)}
\usage{thigh(Close, High = NULL, lag = 5, plot = TRUE, ...) }
\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{lag}{Integer.
Number of lag periods.}
\item{plot}{Logical. If TRUE plot is returned.}
\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


