\name{thigh}
\alias{thigh}
\title{True High oscillator}
\description{Compute True High oscillator (Technical Analysis)}
\usage{thigh(Close, High = NULL, lag = 5, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{lag}{ INTEGER.
Number of lag periods.}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


