\name{track_sign_exp}
\alias{track_sign_exp}
\title{Exponential track signal}
\description{Calculate Exponential track signal for model evaluation }
\usage{track_sign_exp(actual, predicted, beta = 0.1) }
\arguments{\item{actual}{ actual}\item{predicted}{ predicted}\item{beta}{ beta}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
