\name{trf}
\alias{trf}
\title{(Average) True range}
\description{Compute (Average) True range (Technical Analysis)}
\usage{trf(Close, High = NULL, Low = NULL, lag = 1, 
average = TRUE, avg.lag = 14, plot = FALSE, ...)
}
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{lag}{ INTEGER. Number of lag periods.}\item{average}{ average}\item{avg.lag}{
avg.lag}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments to or from
other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
