\name{trf}
\alias{trf}

\title{(Average) True range}

\description{Compute (Average) True range (Technical Analysis)}

\usage{trf(Close, High = NULL, Low = NULL, lag = 1, average = TRUE, avg.lag = 14, plot = FALSE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{lag}{ }\item{average}{ }\item{avg.lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
