\name{ulcer}
\alias{ulcer}
\title{Ulcer index}
\description{Compute Ulcer index (Technical Analysis)}
\usage{ulcer(X, lag, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


