\name{ultima}
\alias{ultima}

\title{Ultima oscillator}

\description{Compute Ultima oscillator (Technical Analysis)}

\usage{ultima(Close, High = NULL, Low = NULL, lag = 1, win1 = 7, win2 = 14, win3 = 28, plot = TRUE, ...) }

\arguments{ \item{Close}{ }\item{High}{ }\item{Low}{ }\item{lag}{ }\item{win1}{ }\item{win2}{ }\item{win3}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
