\name{ultima}
\alias{ultima}
\title{Ultima oscillator}
\description{Compute Ultima oscillator (Technical Analysis)}
\usage{ultima(Close, High = NULL, Low = NULL, lag = 1, win1 = 7, win2 = 14, win3 = 28, plot = TRUE,
...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{lag}{ INTEGER. Number of lag periods.}\item{win1}{ win1}\item{win2}{
win2}\item{win3}{ win3}\item{plot}{ LOGICAL. If TRUE plot is returned.}\item{...}{ Further arguments
to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
