\name{vcmof}
\alias{vcmof}

\title{Variable Chande Momentum Oscillator}

\description{Compute Variable Chande Momentum Oscillator (Technical Analysis)}

\usage{vcmof(X, lag = 5, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
