\name{vhff}
\alias{vhff}
\title{Vertical Horizontal Filter}
\description{Compute Vertical Horizontal Filter (Technical Analysis)}
\usage{vhff(X, lag = 9, plot = FALSE, ...) }
\arguments{
\item{X}{ X}
\item{lag}{Integer. Number of lag periods.}
\item{plot}{Logical. If TRUE
plot is returned.}
\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


