\name{vhff}
\alias{vhff}

\title{Vertical Horizontal Filter}

\description{Compute Vertical Horizontal Filter (Technical Analysis)}

\usage{vhff(X, lag = 9, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
