\name{vidyaf}
\alias{vidyaf}

\title{Variable Index Dynamic Average}

\description{Compute Variable Index Dynamic Average (Technical Analysis)}

\usage{vidyaf(X, lag = 5, plot = FALSE, ...) }

\arguments{ \item{X}{ }\item{lag}{ }\item{plot}{ }\item{...}{ }}

\details{}

\value{}

\references{}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}

\seealso{}

\examples{}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
