\name{vidyaf}
\alias{vidyaf}
\title{Variable Index Dynamic Average}
\description{Compute Variable Index Dynamic Average (Technical Analysis)}
\usage{vidyaf(X, lag = 5, plot = FALSE, ...) }
\arguments{\item{X}{ X}\item{lag}{ INTEGER. Number of lag periods.}\item{plot}{ LOGICAL. If TRUE
plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


