\name{wad}
\alias{wad}
\title{Williams Advance Decline}

\description{Compute Williams Advance Decline (Technical Analysis)}

\usage{
wad(Close
    , High = NULL
    , Low = NULL
    , lag = 5
    , na.rm = FALSE
    , plot = TRUE
    , ...) 
}

\arguments{
\item{Close}{Vector. Close price.}
\item{High}{Vector. High price.}
\item{Low}{Vector.
Low price.}
\item{lag}{Integer. Number of lag periods.}
\item{na.rm}{ na.rm}
\item{plot}{Logical. If
TRUE plot is returned.}
\item{...}{ Further arguments to or from other methods.}}

\author{RAdamant Development Team \email{team@r-adamant.org}}

\note{TO BE COMPLETED}


