\name{wad}
\alias{wad}
\title{Williams Advance Decline}
\description{Compute Williams Advance Decline (Technical Analysis)}
\usage{wad(Close, High = NULL, Low = NULL, lag = 5, na.rm = FALSE, plot = TRUE, ...) }
\arguments{\item{Close}{ VECTOR. Close price.}\item{High}{ VECTOR. High price.}\item{Low}{ VECTOR.
Low price.}\item{lag}{ INTEGER. Number of lag periods.}\item{na.rm}{ na.rm}\item{plot}{ LOGICAL. If
TRUE plot is returned.}\item{...}{ Further arguments to or from other methods.}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}


