\name{wildSum}
\alias{wildSum}
\title{Wilder Summation}
\description{Compute Wilder Summation (Technical Analysis)}
\usage{wildSum(x, lag = 5) }
\arguments{\item{x}{ x}\item{lag}{ lag}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
