\name{Zind}
\alias{Zind}
\title{Z index}
\description{Compute the Z-score of X (Standardize each column of X)}
\usage{Zind(x, sigma = 1, mi = 2) }
\arguments{\item{x}{ x}\item{sigma}{ sigma}\item{mi}{ mi}}
\author{RAdamant Development Team \email{team@r-adamant.org}}
\note{TO BE COMPLETED}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
