\name{Zscore}
\alias{Zscore}
\title{Z Score}
\description{Compute the Z-score of X (Standardize each column of X)}
\usage{Zscore(X, means = NULL, sigma = NULL) }
\arguments{ 
\item{X}{ Matrix of data series (one column per variable)}
\item{means}{ Mean value}
\item{sigma}{ Standard deviation}
}
\value{Matrix of standardised variables}
\author{RAdamant Development Team \email{team@r-adamant.org} }
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
